% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_commutability_evaluation_plots.R
\name{apply_theme_attributes}
\alias{apply_theme_attributes}
\title{Apply Plotting Theme Attributes: PCEP}
\usage{
apply_theme_attributes(plot_theme, additional_arguments)
}
\arguments{
\item{plot_theme}{A \code{character} string.}

\item{additional_arguments}{A \code{list}.}
}
\value{
A \code{list}. The modified \code{additional_arguments}.
}
\description{
Apply Plotting Theme Attributes: PCEP
}
\keyword{internal}
