% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_assessment_plots.R
\name{apply_theme_attributes_pap}
\alias{apply_theme_attributes_pap}
\title{Applies \code{plot_theme} to \code{additional_arguments}}
\usage{
apply_theme_attributes_pap(plot_theme, additional_arguments)
}
\arguments{
\item{plot_theme}{A \code{character} string. The desired plotting theme.}

\item{additional_arguments}{See \code{?plot_assessment_plots()}.}
}
\value{
The original \code{additional_arguments} with the theme attributes
        applied to it.
}
\description{
This function is used to apply the selected \code{plot_theme} to the
\code{additional_arguments} argument in \code{plot_assessment_plots()}.
}
\keyword{internal}
