% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{apply_transformation}
\alias{apply_transformation}
\title{Apply a Univariate Transformation to \code{MP_A} and \code{MP_B} in \code{data}}
\usage{
apply_transformation(data, type, parameter)
}
\arguments{
\item{data}{A \code{data.table}. The IVD-MD comparison results for a
particular IVD-MD comparison.}
}
\value{
A \code{data.table}. The transformed data for a particular IVD-MD
comparison.
}
\description{
This function is internally to apply a univariate transformation to each
numeric column (\code{MP_A} and \code{MP_B}) in \code{data}. Depends on the
last two list elements of the output of
\code{extract_transformation_information()}. Not to be used by end-users.
}
\keyword{internal}
