% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_zeta_data.R
\name{approximate_upper_zeta}
\alias{approximate_upper_zeta}
\title{Internal Function for Approximating Critical Values of \eqn{\hat{\zeta}}}
\usage{
approximate_upper_zeta(data_list, N, M)
}
\description{
This function is used internally for approximating critical values of
\eqn{\hat{\zeta}}, using Monte Carlo simulations. Not meant to be used by
end-users.
}
\keyword{internal}
