% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_commutability_evaluation.R
\name{attempt_fix_upper_zeta}
\alias{attempt_fix_upper_zeta}
\title{Attempt to Fix Upper Zeta}
\usage{
attempt_fix_upper_zeta(upper_zeta, avoid_simulations)
}
\arguments{
\item{upper_zeta}{Object to be checked.}

\item{avoid_simulations}{Object to be checked.}
}
\value{
Returns nothing if check is successful. Otherwise, an error is thrown.
}
\description{
Try to fix upper zeta
}
\keyword{internal}
