% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_assessment_tests.R
\name{bonferroni_to_holm}
\alias{bonferroni_to_holm}
\title{Converts Standard to Holm-adjusted p-values and Test Conclusions}
\usage{
bonferroni_to_holm(data, level, n_sim_tests = 10)
}
\arguments{
\item{data}{See \code{?perform_assessment_tests()}.}

\item{level}{See \code{?perform_assessment_tests()}.}

\item{n_sim_tests}{An \code{integer}. The number of simultanuous tests to
perform for normality and homoscedasticity separately.}
}
\value{
A \code{data.table}, where the p-values and conclusions are modified based
on the Holm-adjustment.
}
\description{
This function modifies Bonferroni-adjusted and non-adjusted p-values
to Bonferroni-Holm-adjusted p-values. The test conclusion is altered based
on this modification too. However, \code{rejection_rate} is not touched.
}
\keyword{internal}
