% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prediction_data.R
\name{bootstrap_inside_rate}
\alias{bootstrap_inside_rate}
\title{Estimate Inside Rate Using Bootstrap Resampling}
\usage{
bootstrap_inside_rate(
  B,
  data,
  new_data,
  R,
  R_ratio,
  level = 0.95,
  method = "fg",
  rounding = 3L
)
}
\arguments{
\item{B}{An \code{integer}. Must be larger than or equal to \code{2}. The
desired number of bootstrap replicates used to estimate the inside
rate.}

\item{data}{A \code{data.table}. Should contain the following variables:
\itemize{
 \item \code{SampleID: } A \code{character} vector. The sample
       identifiers for the clinical samples.
 \item \code{MP_A: } A \code{numeric} vector. The observed
       measurement results from IVD-MD \code{MP_A} (response).
 \item \code{MP_B: } A \code{numeric} vector. The observed
       measurement results from IVD-MD \code{MP_B} (predictor).
}}

\item{new_data}{A \code{data.table}. Should contain the same variables as
\code{data}.}

\item{R}{An \code{integer}. Must be larger than or equal to \code{2}. The
ceiling of the average number of replicated measurements used in
\code{new_data}.}

\item{R_ratio}{A \code{double}. The ratio between the ceiling of the average
number of replicated measurements used in \code{data} and
\code{R}.}

\item{level}{A \code{double}. Must be between \code{0} and \code{1}. The
desired nominal confidence level for the estimated prediction
intervals. Defaults to \code{0.99} (\eqn{99\%}).}

\item{method}{A \code{character} string. The method used to estimate
prediction intervals. See \code{?estimate_prediction_data} for
more information.}

\item{rounding}{An \code{integer}. The number of decimals to include in the
\code{numeric} results. Defaults to \code{3L}.}
}
\value{
A \code{data.table}. Contains the inside rates (denoted by
\code{inside_rate}) for each unique \code{SampleID} in \code{new_data}.
}
\description{
Estimate the \code{SampleID}-wise inside rates by resampling the
\code{inside} indicator variables \code{B} times.
}
\details{
For each \eqn{b = 1, 2, \ldots, B}, resample \code{data}. This results in
\code{B} prediction datasets, which contains the \code{inside} indicator
variables. For each unique \code{SampleID} in \code{new_data}, we take the
average of the \code{B} resampled values of \code{inside} and this is the
resulting estimated inside rate.

Note that the estimation of inside rates using either \code{method = 'ss'}
or \code{method = 'ssw'} will take considerably longer than using either
of the linear choices of \code{method}.
}
\examples{
# Required packages
library(fasteqa)

test_cs_data <- test_data[SampleID != "1" & SampleID != "2"]
test_eq_data <- test_data[SampleID == "1" | SampleID == 2, fun_of_replicates(.SD)]

print(bootstrap_inside_rate(B = 20,
                            data = test_cs_data,
                            new_data = test_eq_data,
                            R = 3,
                            R_ratio = 1,
                            level = 0.95,
                            method = "fg",
                            rounding = 3L))


}
