% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button_style.R
\name{button_style}
\alias{button_style}
\title{Add CSS style for buttons}
\usage{
button_style(event = "default", button_type = "browse")
}
\arguments{
\item{event}{\code{character} - What pseudo class or combination of pseudo classes referring to the button. Some choices are 'default' (no pseudo class), 'hover' and 'active'}

\item{button_type}{\code{character} - What type of button are we talking about. Possible choices are 'browse' and 'radio'.}
}
\value{
\code{character} that contains appropriate CSS style commands for the buttons relative pseudo class
}
\description{
Add CSS style for buttons
}
\details{
This function has no immediate use to end-users. It is just part of the commutability application simplifying the script somewhat relative to tidiness
}
\examples{
button_style('default', 'radio')
}
