% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{calculate_data_quality_score}
\alias{calculate_data_quality_score}
\title{Calculate Data Quality Score of \code{data}}
\usage{
calculate_data_quality_score(
  effective_number_of_samples,
  influence_number_of_samples,
  average_number_of_replicates,
  fraction_of_NAs,
  type = "cs"
)
}
\arguments{
\item{effective_number_of_samples}{An \code{integer} vector. Number of
effective samples for each column.}

\item{influence_number_of_samples}{A \code{numeric} vector. Relative
influence of each column's sample size.}

\item{average_number_of_replicates}{A \code{numeric} vector. Average number
of replicates for each column.}

\item{fraction_of_NAs}{A \code{list}. Fraction of NAs for each column.}

\item{type}{A \code{character} string. Either \code{'cs'} for clinical
samples or \code{'eqam'} for evaluated materials.}
}
\value{
An \code{integer}. The quality score. The calculated quality score (0-9).
}
\description{
This function computes a quality score for the dataset based on various
metrics such as sample size, replicate numbers, and missing data.
}
\details{
The function uses a point system based on four main criteria:
\itemize{
 \item Minimum number of effective clinical samples
 \item Minimum average number of replicates
 \item Similarity of effective clinical samples across columns
 \item Largest NA fraction among columns
}
}
\keyword{internal}
