% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_equivalence.R
\name{check_equivalence}
\alias{check_equivalence}
\title{Checking for equivalent forms of clinical sample data and external quality assessment material data}
\usage{
check_equivalence(cs_data, eq_data, silence = 1L)
}
\arguments{
\item{cs_data}{data table, data frame or list - dataset containing clinical samples' measurements.}

\item{eq_data}{data table, data frame or list - dataset containing external quality assessment material samples' measurements}

\item{silence}{integer - Should verbose be allowed? Additional potentially helpful text are printed if silence is set to smaller than 1L}
}
\value{
A list with three elements with information regarding equivalence status
}
\description{
Checking for equivalent forms of clinical sample data and external quality assessment material data
}
\examples{
print(1)
}
