% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_zeta_data.R
\name{check_estimate_zeta_data}
\alias{check_estimate_zeta_data}
\title{Internal Function for Checking Input Data}
\usage{
check_estimate_zeta_data(
  data,
  B = 2000L,
  type = "percentile",
  level = 0.95,
  M = 0,
  N = 10000L,
  zeta_critical = NA,
  method = c("ols", "ss", "ssw"),
  invalid_NA = FALSE
)
}
\description{
This function is used internally for checking and possibly converting the
input of \code{estimate_zeta_data} to correct format.
}
\keyword{internal}
