% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_mandatory_id_columns}
\alias{check_mandatory_id_columns}
\title{Check for Mandatory ID Columns in \code{data}}
\usage{
check_mandatory_id_columns(data)
}
\arguments{
\item{data}{A \code{data.table} to be checked.}
}
\value{
A \code{list} containing:
\itemize{
 \item \code{pass: } A \code{logical} value. Indicates whether both
       mandatory ID columns were found.
 \item \code{data: } The input \code{data} with standardized column names
       for \code{SampleID} and \code{ReplicateID}, if found.
}
}
\description{
This function checks for the presence of mandatory ID columns
(\code{SampleID} and \code{ReplicateID}) in the input \code{data}. It
attempts to identify these columns even if they have slight naming
variations.
}
\keyword{internal}
