% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_na_fraction}
\alias{check_na_fraction}
\title{Check Validity of Number of Missing Values in \code{data}}
\usage{
check_na_fraction(data, type = "cs")
}
\arguments{
\item{data}{A \code{data.table} containing the data to be checked.}

\item{type}{A \code{character} string. Either \code{'cs'} for clinical
samples or \code{'eqam'} for evaluated materials.}
}
\value{
A \code{list} containing:
\itemize{
 \item \code{pass: } A \code{logical} indicating if the data passes the
       \code{NA} fraction check.
 \item \code{number_of_NAs_numeric_data: } Number of NAs for each
       \code{numeric} column.
 \item \code{fraction_of_NAs_numeric_data: } Fraction of NAs for each
       \code{numeric} column.
 \item \code{amount_of_NAs_numeric_data: } Categorical classification of NA
       amount.
 \item \code{effective_numbers_of_samples: } Number of non-NA samples for
       each column.
 \item \code{average_number_of_replicates: } Average number of replicates
       per sample.
 \item \code{influence_number_of_samples: } Relative influence of each
       column's sample size.
 \item \code{minimum_number_of_samples: } Minimum effective number of
       samples across columns.
 \item \code{remove_these_methods: } Columns that should be removed due to
       low quality.
 \item \code{keep_these_ID_column_ids: } IDs of rows with valid SampleID and
       ReplicateID
 \item \code{keep_these_numeric_column_ids: } IDs of rows with valid numeric
       data after excluding invalid SampleID and ReplicateID rows.
}
}
\description{
This function assesses the number of missing data points in the input
\code{data}, calculating various metrics related to sample size and data
completeness.
}
\keyword{internal}
