% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prediction_data.R
\name{check_names_epd}
\alias{check_names_epd}
\title{Validates Names of \code{data} and \code{new_data}}
\usage{
check_names_epd(data_list, new_data_list)
}
\arguments{
\item{data_list}{Object to be checked.}

\item{new_data_list}{Object to be checked.}
}
\value{
Returns nothing if check is successful. Otherwise, an error is thrown.
}
\description{
This function is used to check if \code{data} and \code{new_data} have the
same names and the same order. This is a helper function used in
\code{estimate_prediction_data}
}
\keyword{internal}
