% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_numeric_columns}
\alias{check_numeric_columns}
\title{Check Validity of Numeric Columns in \code{data}}
\usage{
check_numeric_columns(data)
}
\arguments{
\item{data}{A \code{data.table} containing the data to be checked.

A \code{list} containing:
\itemize{
 \item \code{pass: } A \code{logical} value. Indicates whether the
       \code{numeric} columns meet the validity criteria.
 \item \code{numeric_columns: } A \code{character} vector of column names
       identified as \code{numeric}.
 \item \code{convert_these_numeric_columns: } A \code{character} vector of
       column names that should be converted to numeric.
}}
}
\description{
This function assesses the \code{numeric} columns in the input \code{data},
checking for their validity and potential for conversion to \code{numeric}
format if they are not already.
}
\details{
The function considers a column "numeric enough" if at least 90% of its
non-NA values can be converted to \code{numeric}.
}
\keyword{internal}
