% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_look_up_table.R
\docType{data}
\name{commutability_cs_data}
\alias{commutability_cs_data}
\title{Example Clinical Sample Data}
\format{
A \code{data.table} with 123 rows and 12 variables:
\describe{
  \item{SampleID}{An \code{integer} vector. The clinical sample identifiers}
  \item{ReplicateID}{A \code{character} vector. The replicate measurement identifiers}
  \item{other}{IVD-MD measurements}
}
}
\usage{
commutability_cs_data
}
\description{
An example dataset containing clinical sample measurements for ten unique
IVD-MDs. \code{Thermo} is considered a reference IVD-MD.
}
\keyword{datasets}
