% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_look_up_table.R
\docType{data}
\name{commutability_eq_data}
\alias{commutability_eq_data}
\title{Example Evaluated Material Sample Data}
\format{
A \code{data.table} with 183 rows and 12 variables:
\describe{
  \item{SampleID}{An \code{character} vector. The evaluated material sample identifiers}
  \item{ReplicateID}{A \code{character} vector. The replicate measurement identifiers}
  \item{other}{IVD-MD measurements}
}
}
\usage{
commutability_eq_data
}
\description{
An example dataset containing evaluated material sample measurements for
ten unique IVD-MDs. This dataset contains data for \eqn{61} evaluated
materials. \code{Thermo} is considered a reference IVD-MD.
}
\keyword{datasets}
