% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{convert_data_cd}
\alias{convert_data_cd}
\title{Convert \code{data} to \code{data.table} Format}
\usage{
convert_data_cd(data)
}
\arguments{
\item{data}{A \code{data.table}, \code{list}, or \code{data.frame} to be
converted.}
}
\value{
A \code{list} containing:
\itemize{
 \item \code{data: } The converted \code{data.table} object, or the original
       input if conversion failed.
 \item \code{can_convert_data: } A \code{logical} value. Indicates whether
       the conversion was successful.
}
}
\description{
This function attempts to convert the input \code{data} to a
\code{data.table} format. It handles \code{data.frame} and \code{list}
inputs, converting them to \code{data.table} if possible.
}
\keyword{internal}
