% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_style.R
\name{css_style}
\alias{css_style}
\title{Add CSS style for other R shiny elements}
\usage{
css_style(event = "default", element = "input_field")
}
\arguments{
\item{event}{\code{character} - What pseudo class or combination of pseudo classes referring to the element. Some choices are 'default' (no pseudo class), 'hover' and 'active'}

\item{element}{\code{character} - What type of element are we talking about. Some possible choices are 'well' and 'input_field' among others.}
}
\value{
\code{character} that contains appropriate CSS style commands for the given element's relative pseudo class
}
\description{
Add CSS style for other R shiny elements
}
\details{
This function has no immediate use to end-users. It is just part of the commutability application simplifying the script somewhat relative to tidiness
}
\examples{
css_style('hover', 'progress_bar')
}
