% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_commutability_evaluation.R
\name{do_commutability_evaluation}
\alias{do_commutability_evaluation}
\title{Run a Complete Commutability Evaluation Analysis Based on Clinical Sample
Data and Evaluated Material Data}
\usage{
do_commutability_evaluation(
  data,
  new_data,
  B = 2000L,
  N = 1000L,
  method_pi = "fg",
  method_bs = "percentile",
  level_pi = 0.99,
  level_bs = 0.95,
  M = 0,
  upper_zeta = 2.25,
  output = "sufficient",
  avoid_simulations = FALSE
)
}
\arguments{
\item{data}{A \code{list}, \code{data.table} or \code{data.frame}. Must
contain the measurements of clinical samples grouped by
\code{comparison}, \code{SampleID} and \code{ReplicateID}.}

\item{new_data}{A \code{list}, \code{data.table}, or \code{data.frame}. Must
contain the measurements of either external quality
assessment materials or reference materials grouped by
\code{comparison}, \code{SampleID} and \code{ReplicateID}.}

\item{B}{An \code{integer} \eqn{geq} 100 that represents the number of bootstrap
replicates utilized for estimating bootstrap confidence intervals.
The default is set to \code{2000}, a common choice for such
calculations. Please be aware that if your \code{data} contains
more than five unique IVD-MDs, the resampling process might take a
couple of seconds to complete due to the computational complexity.}

\item{N}{An \code{integer} \eqn{geq} 100 that specifies the number of
bootstrap resamples used for estimating the inside rates
(commutability conclusion consistency) for the evaluated material.
Set to \code{NULL} to exclude estimation of inside rates.}

\item{method_pi}{A \code{character} string referring to the method to employ
for estimating the Deming prediction intervals. The default
is \code{'fg'}, standing for the Fuller and Gillard
approach. Alternatively, the approach derived by CLSI EP14
can be used by setting \code{'clsi'}.}

\item{method_bs}{A \code{character} string. Determines the bootstrap
method to apply for estimating the confidence intervals
of \eqn{\zeta} and repeatability statistics for each IVD-MD
comparison.}

\item{level_pi}{A \code{double}. Must be between \code{0} and \code{1}.
This value is the desired nominal confidence level for the
estimated prediction intervals. The default is \code{0.99}
(\eqn{99\%}).}

\item{level_bs}{A \code{double}. Must be between 0 and 1. The desired
nominal confidence level for the bootstrap estimated
confidence intervals. The default is \code{0.95}
(\eqn{95\%}).}

\item{M}{A \code{double} \eqn{\geq 0}. If \code{upper_zeta} is either
\code{NULL}, \code{NA}, or an improperly formatted \code{character}
string (see details), and \code{avoid_simulations = FALSE}, this
parameter will signify the acceptable level of differences in
non-selectivity between all IVD-MD pairs within \code{data} and
\code{new_data}. This tolerance level is expressed as an average
relative increase in the widths of pointwise prediction intervals.}

\item{upper_zeta}{A \code{double} \eqn{\geq 0}. The maximum value \eqn{\hat{\zeta}}
can take that correspond with an acceptable magnitude of
differences in non-selectivity.}

\item{output}{A \code{character} string. The output structure:
\itemize{
   \item \code{'sufficient': } Only return essential results.
   \item \code{'complete': } Return all results.
}}

\item{avoid_simulations}{A \code{logical} value. Ensures that an error is
thrown if \code{upper_zeta} is a \code{character}
string with invalid syntax (see details) instead of
approximating \code{upper_zeta} values utilizing
tedious Monte Carlo simulations. Defaults to
\code{FALSE}.}
}
\value{
A \code{list} of length two. Contains:
\itemize{
   \item \code{merged_pb_data: } Comprises pointwise prediction intervals
         (prediction band data). Intended solely for plotting purposes
   \item \code{merged_ce_data: } A thorough, tabulated summary of the
         commutability evaluation data analysis.
}
}
\description{
Run a Complete Commutability Evaluation Analysis Based on Clinical Sample
Data and Evaluated Material Data
}
\examples{
print(1)
}
