% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_outlier_analysis.R
\name{do_outlier_analysis}
\alias{do_outlier_analysis}
\title{Run a complete outlier analysis based on CS data}
\usage{
do_outlier_analysis(
  data,
  type = c("burnett", "qrange", "3sd", "5cv", "iqr"),
  where = c("replicates", "samples"),
  what = c("raw", "rd", "ad", "ld"),
  simplify = FALSE,
  silence = 1L,
  testing = FALSE
)
}
\arguments{
\item{data}{\code{list}, \code{data table} or \code{data frame} - Clinical sample data (CS data) for the commutability assessment experiment}

\item{type}{\code{character} - Which type of outlier assessment test should be applied to the input data? Possible choices are \code{'burnett'} (default), \code{'qrange'}, \code{'3sd'} and \code{'5cv'}.}

\item{where}{\code{character} - Where should the function look for outliers? Possible values are \code{'replicates'} (default) and \code{'samples'}. The former searches for outliers within replicated measurements, whereas the latter searces for outliers between samples.}

\item{what}{\code{character} - What transformation should be applied before looking for outliers? Possible choices are \code{'raw'} (default), \code{'rd'}, \code{'ad'} and \code{'ld'}. The three latter choices are shorthand for relative difference, absolute difference and log difference, respectively.}

\item{simplify}{\code{logical} - Should the output be attempted to be simplified? If \code{TRUE}, the output will be a \code{data.table} object with 'comparison' as the ID column. In contrast, if \code{FALSE}, a list is returned.}

\item{silence}{\code{integer} - Should verbose and debugging messages be included in output? Set \code{silence > 0L} to suppress such messages.}

\item{testing}{\code{unknown} - Only for maintainer - Should the function provide test appropriate outputs. Default is \code{FALSE}.}
}
\value{
a \code{data.table} object that lists the outlier analysis results. The dimensions of the output depends on \code{type}, \code{where} and \code{what}.
}
\description{
Run a complete outlier analysis based on CS data
}
\examples{
print(1)
}
