% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_imprecision_data.R
\name{estimate_imp_boot_ci}
\alias{estimate_imp_boot_ci}
\title{Estimates Imprecision Bootstrap Confidence Intervals}
\usage{
estimate_imp_boot_ci(B, data, type, level)
}
\arguments{
\item{B}{An \code{integer}. The desired number of bootstrap replicates used
to estimate bootstrap confidence intervals. See
\code{?estimate_imprecision_data()} for more information.}

\item{data}{A \code{data.table} or \code{list}. The clinical sample data.
See \code{?estimate_imprecision_data()} for more information.}

\item{type}{An \code{integer}. The type of bootstrap confidence interval
desired. Can be \code{1-4}.}

\item{level}{A \code{double}. The desired nominal confidence level of the
estimated boostrap confidence intervals.}
}
\value{
A \code{list}. The estimated boostrap confidence intervals.
}
\description{
Estimates Imprecision Bootstrap Confidence Intervals
}
\details{
Not to be used by end-users. Used internally in
\code{estimate_imprecision_data()}.
}
\keyword{internal}
