% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prediction_data.R
\name{estimate_prediction_data}
\alias{estimate_prediction_data}
\title{Estimate Prediction Intervals Based on Clinical Sample (CS) data and
Evaluated Material Data}
\usage{
estimate_prediction_data(
  data,
  new_data = NULL,
  B = NULL,
  method = "fg",
  level = 0.99,
  rounding = 3L,
  override_R_ratio = NULL,
  na_rm = TRUE
)
}
\arguments{
\item{data}{A \code{list}, \code{data.table} or \code{data.frame}. The
CS data. Must contain the following variables:
\itemize{
   \item \code{comparison: } A \code{character} vector. The
         comparison identifiers. Typically on the form
         \code{'MP_A - MP_B'}.
   \item \code{SampleID: } A \code{character} vector. The sample
         identifiers for the clinical samples.
   \item \code{ReplicateID: }A \code{character} vector. The
         replicated measurement identifiers.
   \item \code{MP_A: } A \code{numeric} vector. The observed
         measurement results from IVD-MD \code{MP_A} (response).
   \item \code{MP_B: } A \code{numeric} vector. The observed
         measurement results from IVD-MD \code{MP_A} (predictor).
}}

\item{new_data}{A \code{list}, \code{data.table} or \code{data.frame}. The
CS data. Must contain the following variables:
\itemize{
   \item \code{comparison: } See \code{data}.
   \item \code{MP_B: } See \code{data}.
}
Can contain:
\itemize{
   \item \code{SampleID: } A \code{character} vector. The
   sample identifiers for the evaluated material samples.
   \item \code{ReplicateID: } See \code{data}.
   \item \code{MP_A: } See \code{data}.
}
Can alternatively be a \code{character} string. See details.}

\item{B}{An \code{integer}. The desired number of bootstrap resamples to be
used for estimating the inside rates (commutability conclusion
consistency) for the evaluated materials. If \code{NULL} (default),
inside rates are not calculated.}

\item{method}{A \code{character} string. The method used to estimate
              prediction intervals. The default is \code{'fg'}. Current
              possible prediction interval estimation methods include:
\itemize{
  \item \code{fg: } Implements standard Deming regression with prediction
         intervals estimated using the Fuller and Gillard method.
  \item \code{clsi: } Utilizes standard Deming regression with prediction
                      intervals estimated using the CLSI EP14 approach.
  \item \code{ols: } Implements Ordinary Least Squares regression.
  \item \code{ss: } Implements Smoothing Splines with prediction intervals
                    approximated using an heuristic extension of the
                    bayesian confidence intervals derived by Grace Wahba.
                    See references.
  \item \code{ssw: } Implements Weighted Smoothing Splines with prediction
                     intervals approximated using a heuristic extension of
                     the bayesian confidence intervals derived by Grace
                     Wahba. See references.
}}

\item{level}{A \code{double}. Must be between \code{0} and \code{1}. The
desired nominal confidence level for the estimated prediction
intervals. Defaults to \code{0.99} (\eqn{99\%}).}

\item{rounding}{An \code{integer}. The number of decimals to include in the
\code{numeric} results. Defaults to \code{3L}.}

\item{override_R_ratio}{A \code{list} or \code{NULL} (default). When
provided as a \code{list}, it should include the
IVD-MD comparison-specific ratios of the number of
replicates between evaluated material data and CS
data.}

\item{na_rm}{A \code{logical} value. Indicates whether \code{NA} values are
silently removed from \code{data} and \code{new_data}. Defaults
to \code{TRUE}.}
}
\value{
A \code{data.table}. Depending on the structure of \code{new_data},
contains either commutability evaluation data (CE data) or prediction band
data (PB data).
}
\description{
Estimate Prediction Intervals Based on Clinical Sample (CS) data and
Evaluated Material Data
}
\details{
The output of this function depends greatly on the input of \code{new_data}.
If \code{new_data} contains \code{SampleID} and \code{MP_A}, the function
returns the so-called CE (Commutability Evaluation) data. If either of these
are omitted, or if \code{new_data} is a \code{character} string, the
function returns the so-called PB (Prediction Band) data.

Note: CE data is typically coded \code{ce_data}. PB data is often coded
\code{pb_data}. These codings are important to remember because they are
used in e.g., \code{do_commutability_evaluation()} and
\code{plot_commutability_evaluation_plots()}.

When \code{new_data} is a \code{character} string it most be on the format
\code{'gen#number'} or \code{'number#gen'}, where \code{number} is an
integer. For example, setting \code{new_data = 'gen#100'} or
\code{new_data = '100#gen'} would instruct the function to generate
\code{100} pointwise prediction intervals uniformly distributed across the
domain of the IVD-MD measurements for each comparison.
}
\examples{
# Required packages
library(smooth.commutability)
library(fasteqa)
library(data.table)

test_crp_cs_data <- copy(crp_cs_data)
test_crp_eq_data <- copy(crp_eqam_data)

# Estimate CE (Commutability Evaluation) data using weighted smoothing spline
print(estimate_prediction_data(data = test_crp_cs_data,
                               new_data = test_crp_eq_data[SampleID == "1"],
                               B = 50,
                               method = "ssw",
                               level = 0.99,
                               rounding = 3L,
                               override_R_ratio = NULL,
                               na_rm = TRUE))

# Estimate PB (Prediction Band) data using weighted smoothing spline
pb_data <- estimate_prediction_data(data = test_crp_cs_data,
                                    new_data = "gen#20",
                                    B = NULL,
                                    method = "ssw",
                                    level = 0.99,
                                    rounding = 2L,
                                    override_R_ratio = NULL,
                                    na_rm = TRUE)

# The PB data for the AQT90 - Chroma comparison
pb_data_AQT90_Chroma <- pb_data[comparison == "AQT90 - Chroma"]
print(pb_data_AQT90_Chroma)

# Plotting the PB data for the AQT90 - Chroma comparison
mor_test_crp_cs_data <- test_crp_cs_data[comparison == "AQT90 - Chroma", fun_of_replicates(.SD)]
plot(x = mor_test_crp_cs_data$MP_B,
     y = mor_test_crp_cs_data$MP_A,
     main = "Prediction Band Data For AQT90 - Chroma",
     xlab = "Chroma",
     ylab = "AQT90",
     pch = 20)
lines(x = pb_data_AQT90_Chroma$predictor,
      y = pb_data_AQT90_Chroma$lwr,
      col = "red")
lines(x = pb_data_AQT90_Chroma$predictor,
      y = pb_data_AQT90_Chroma$upr,
      col = "red")

}
