% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_outlier_analysis.R
\name{estimate_variable}
\alias{estimate_variable}
\title{Calculate the Desired \code{variable}}
\usage{
estimate_variable(data, variable)
}
\arguments{
\item{data}{A \code{data.table}.}

\item{variable}{A \code{character} string. See \code{?do_outlier_analysis()}.}
}
\value{
The same as \code{data}, but where \code{variable} is calculated based on
\code{MP_A} and \code{MP_B} for each \code{SampleID} in \code{data}.
}
\description{
This function is used to calculate the desired \code{variable} to be used in
\code{do_outlier_analysis()}. This function should not be used by end-users.
}
\keyword{internal}
