% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_zeta_data.R
\name{estimate_zeta_boot_ci}
\alias{estimate_zeta_boot_ci}
\title{Internal Function for Calculating Boostrap Confidence Intervals}
\usage{
estimate_zeta_boot_ci(
  data_list,
  original_zetas,
  original_dfs = NULL,
  B = B,
  method,
  type,
  level,
  buffer = 0.05
)
}
\description{
This function is used internally for Calculating Boostrap Confidence
Intervals in \code{estimate_zeta_data}
}
\keyword{internal}
