% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_zeta_data.R
\name{estimate_zeta_data}
\alias{estimate_zeta_data}
\title{All calculations regarding zeta}
\usage{
estimate_zeta_data(
  data,
  B = 2000,
  type = "percentile",
  level = 0.95,
  M = 0,
  zeta_critical = NA
)
}
\arguments{
\item{data}{\code{list} or \code{data table} grouped by \code{"comparison"}. Must contain \code{comparison}, \code{SampleID}, \code{ReplicateID}, \code{MP_A} and \code{MP_B}}

\item{B}{Number of bootstrap replicates used to estimate bootstrap confidence intervals. The default is 2000, which is the typical for bootstrap confidence intervals}

\item{type}{Type of bootstrap confidence interval. There are four options:
\itemize{
  \item{\code{normal}: }{Standard normal bootstrap confidence interval}
  \item{\code{basic}: }{Basic bootstrap confidence interval}
  \item{\code{percentile}: }{Percentile bootstrap confidence interval}
  \item{\code{BCa}: }{Bias- and skewness-corrected bootstrap confidence interval using Jack knife}
}}

\item{level}{Confidence level of bootstrap confidence interval. A 95 percent confidence level is the default}

\item{M}{If \code{zeta_critical} is not specified (i.e., NA or not a double), what \code{M} should be used to obtain a upper value of zeta}

\item{zeta_critical}{A number. Default is NA, which will initiate a Monte Carlo simulation based on your average study design across all MS comparison to obtain an adequate upper zeta value. \code{M} must be specified to obtain a valid upper limit}
}
\value{
A \code{data table} with entries \code{comparison}, \code{zeta}, \code{lwr}, \code{upr}, \code{zeta_critical} and \code{zeta_conclusion}
}
\description{
Obtain all necessary information on zeta data for each unique pair of IVD-MDs in \code{data}. The output is also on the form required by merge_results(), making it very useful
}
\examples{
\dontrun{
  print(1)
}
}
