% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prediction_data.R
\name{extract_n_gens}
\alias{extract_n_gens}
\title{Extracts Number of Pointwise Prediction Intervals From \code{new_data}}
\usage{
extract_n_gens(x)
}
\arguments{
\item{x}{A \code{character} string.}
}
\value{
An \code{integer}.
}
\description{
If \code{new_data} is a \code{character} string on the correct format,
extracts the number of pointwise prediction intervals used to construct
\code{PB} data. This is a helper function used in
\code{estimate_prediction_data}. Should not be used by end-users.
}
\keyword{internal}
