% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{extract_transformation_information}
\alias{extract_transformation_information}
\title{Extract Relevant Information From the \code{transformation} Argument}
\usage{
extract_transformation_information(transformation)
}
\arguments{
\item{transformation}{}
}
\value{
A \code{list} of length four. Relevant extract information from the
character string \code{transformation}.
}
\description{
This function is used to validate and extract relevant information from the
\code{transformation} argument. Not to be used by end-users.
}
\keyword{internal}
