% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prediction_data.R
\name{gen_grid}
\alias{gen_grid}
\title{Generates \code{n} Uniformly Distributed \code{MP_B} Values}
\usage{
gen_grid(n = 100, lwr, upr)
}
\arguments{
\item{n}{An \code{integer}.}

\item{lwr}{A \code{double}. Lower limit for generated values of \code{MP_B}.}

\item{upr}{A \code{double}. Upper limit for generated values of \code{MP_B}.}
}
\value{
A \code{numeric} vector of length \code{n}.
}
\description{
Generates \code{n} uniformly distributed \code{MP_B} values in the interval
\eqn{[\mathrm{lwr}, \mathrm{upr}]}. Should not be used by end-users.
}
\keyword{internal}
