% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_assessment_plots.R
\name{get_assessment_data}
\alias{get_assessment_data}
\title{Get Plotting Data Estimated from \code{data}}
\usage{
get_assessment_data(data, method, type, cv_percent = TRUE, var_instead = FALSE)
}
\arguments{
\item{data}{See \code{?plot_assessment_plots()}.}

\item{method}{See \code{?plot_assessment_plots()}.}

\item{type}{See \code{?plot_assessment_plots()}.}

\item{cv_percent}{A \code{logical} value. If \code{TRUE},
coefficients of variation is presented as percentages.}

\item{var_instead}{A \code{logical} value. If \code{TRUE},
variances are used instead of standard deviations.}
}
\value{
A \code{list} of length \code{2L}:
\itemize{
   \item \code{assessment_data: } A \code{data.table}. Contains the plotting
         data for the desired \code{type}.
   \item \code{supplement_assessment_information: } A \code{data.table} or
         \code{list} that contains additional information that could be
         utilized in the plotting.
}
}
\description{
This function is used to calculate the assessment data used in the plots.
}
\keyword{internal}
