% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_wise.R
\name{get_combos}
\alias{get_combos}
\title{Get Unique IVD-MD Combinations}
\usage{
get_combos(numeric_columns, reference = NULL)
}
\arguments{
\item{numeric_columns}{A \code{character} vector. The names of the IVD-MDs
in the original dataset. Should ideally be sorted
in alphabetical order.}

\item{reference}{A \code{character} string. If not \code{NULL}, the name of
the IVD-MD among \code{numeric_columns} that should be used
as a reference IVD-MD. See details.}
}
\value{
A \code{character} matrix.
}
\description{
Get IVD-MD combinations among the IVD-MDs in \code{numeric_columns}.
}
\details{
If \code{reference} is a \code{character} string among
\code{numeric_columns}, all other IVD-MDs in \code{numeric_columns} are
paired with the IVD-MD provided in \code{reference}. Otherwise, all unique
pairwise combinations of \code{numeric_columns} are used.

This function should not be used by end-users directly. It is used
internally within the \code{get_comparison_data} function.
}
\keyword{internal}
