% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_assessment_tests.R
\name{get_heteroscedasticity_test_results}
\alias{get_heteroscedasticity_test_results}
\title{Execute Heteroscedasticity Tests for \code{data}}
\usage{
get_heteroscedasticity_test_results(
  data,
  level = 0.95,
  B = NULL,
  koenker = TRUE,
  n_sim_tests = 10,
  buffer = 0.05
)
}
\arguments{
\item{data}{See \code{?perform_assessment_tests()}.}

\item{level}{See \code{?perform_assessment_tests()}.}

\item{B}{See \code{?perform_assessment_tests()}.}

\item{koenker}{See \code{robust} in \code{?perform_assessment_tests()}.}

\item{n_sim_tests}{An \code{integer}. The number of simultanuous tests to
perform for normality and homoscedasticity separately.}

\item{buffer}{A \code{double}. Must be non-negative. The percentage increase
in effective \code{B}. May be relevant if the inner
algorithm is prone to producing NAs.}
}
\value{
A \code{data.table}. The Breusch-Pagan heteroscedasticity test results.
}
\description{
This function is used to perform the heteroscedasticity tests that is
part of the output of \code{perform_assessment_tests}.
}
\keyword{internal}
