% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_commutability_evaluation_plots.R
\name{modify_ce_variables_pcep}
\alias{modify_ce_variables_pcep}
\title{Modifies Commutability Evaluation Data Variables}
\usage{
modify_ce_variables_pcep(ce_data)
}
\arguments{
\item{ce_data}{A \code{data.table}.}
}
\value{
A \code{data.table}. The \code{ce_data} processed and ready for used in
\code{plot_commutability_evaluation_plots()}.
}
\description{
Prepare Commutability Evaluaton (CE) data before plotting.
}
\keyword{internal}
