% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_commutability_evaluation_plots.R
\name{modify_pb_variables_pcep}
\alias{modify_pb_variables_pcep}
\title{Modifies Prediction Band Data Variables}
\usage{
modify_pb_variables_pcep(pb_data)
}
\arguments{
\item{pb_data}{A \code{data.table}.}
}
\value{
A \code{data.table}. The \code{pb_data} processed and ready for used in
\code{plot_commutability_evaluation_plots()}.
}
\description{
Prepare Prediction Band (PB) data before plotting.
}
\keyword{internal}
