% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_outlier_analysis.R
\name{outlier_tests}
\alias{outlier_tests}
\title{Calculations and Summaries of Outlier Tests}
\usage{
outlier_tests(data, method, variable, level = 0.99)
}
\arguments{
\item{data}{A \code{data.table}. See \code{?do_outlier_analysis()}.}

\item{method}{A \code{character} string. See \code{?do_outlier_analysis()}.}

\item{variable}{A \code{character} string. See \code{?do_outlier_analysis()}.}

\item{level}{A \code{double} value. See \code{?do_outlier_analysis()}.}
}
\value{
The same as \code{data}, but where \code{variable} is calculated based on
\code{MP_A} and \code{MP_B} for each \code{SampleID} in \code{data}.
}
\description{
This function is used within \code{do_outlier_analysis()} to calculate
test statistics, rejection regions, and test summaries. This function should
not be used by end-users.
}
\keyword{internal}
