% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_confidence_levels_lm.R
\name{parameter_checks_lm}
\alias{parameter_checks_lm}
\title{Internal Function for Checking Input Parameters}
\usage{
parameter_checks_lm(
  N,
  n,
  m,
  level,
  method,
  attach,
  global_mean,
  all_inside,
  calculate_sem
)
}
\description{
This function is used internally for checking if the inputs are valid
}
\keyword{internal}
