% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_assessment_tests.R
\name{perform_assessment_tests}
\alias{perform_assessment_tests}
\title{Perform Linear Model Assessment Tests}
\usage{
perform_assessment_tests(
  data,
  B = 200L,
  method = "fg",
  robust = TRUE,
  level = 0.95,
  adjust_p_value = TRUE,
  holm = TRUE,
  output = "visual"
)
}
\arguments{
\item{data}{A \code{data.table}, \code{list}, or \code{data.frame}.
Must contain the clinical sample data with variables:
\itemize{
   \item \code{comparison: } A \code{character} vector. The
         comparison identifiers. Typically on the form
         \code{'MP_A - MP_B'}.
   \item \code{SampleID: } A \code{character} vector. The sample
         identifiers for the clinical samples.
   \item \code{ReplicateID: } A \code{character} vector. The
         replicated measurement identifiers.
   \item \code{MP_A: } A \code{numeric} vector. The observed
         measurement results from IVD-MD \code{MP_A} (response).
   \item \code{MP_B: } A \code{numeric} vector. The observed
         measurement results from IVD-MD \code{MP_A} (predictor).
}}

\item{B}{An \code{integer}. Must be larger than \code{10}. The number of
bootstrap replicates used to estimate the rejection rate for each
test. See details.}

\item{method}{A \code{character} string. The desired model for calculating
residuals that is used in the Shapiro-Wilk normality test. The
possible options are \code{'fg'} (default), \code{'clsi'},
\code{'ols'}, \code{'ss'} and \code{'ssw'}. The Breusch-Pagan
tests will always use \code{'ols'}, so take care interpreting
the Breusch-Pagan test results under non-linearity.}

\item{robust}{A \code{logical} value. If \code{TRUE} (default), the Koenker
version of the Breusch-Pagan test is used for testing for
variance inhomogeneity.}

\item{level}{A \code{double}. Must be between \code{0} and \code{1}. The
corresponding non-adjusted test significance levels are given
by \eqn{\alpha = 1 - \mathrm{level}}.}

\item{adjust_p_value}{A \code{logical} value. If \code{TRUE} (default),
non-adjusted p-values are Bonferroni-adjusted to
keep the familiy-wise significance level equal to
\eqn{1 - \mathrm{level}} for each test.}

\item{holm}{A \code{logical} value. If \code{TRUE}, non-adjusted p-values
are Holm-Bonferroni corrected. Otherwise (default), the
non-adjusted p-values are Bonferroni corrected. Only relevant
if \code{adjust_p_value} is \code{TRUE}. Note:
\code{rejection_rate} will not account for Holm-Bonferroni
correction.}

\item{output}{A \code{character} string. If \code{'visual'} (default), the
output is formatted in a clean way. However, it is not
optimal to use for further calculations.}
}
\value{
A \code{data.table} with columns:
\itemize{
 \item \code{comparison: } See \code{data}.
 \item \code{test_name: } A \code{character} vector. The test names.
 \item \code{p.value: } A \code{numeric} vector. The raw or adjusted
       p-values.
 \item \code{conclusion: } A \code{character} vector. The conclusion based
       on each test.
 \item \code{rejection_rate: } A \code{numeric} vector. The proportion of
       resampled test statistics that resulted in rejected of the null.
       Will only appear if \code{B} is an integer larger than \code{10}.
}
}
\description{
Perform normality and homoscedasticity tests for each IVD-MD comparison
in \code{data}.
}
\examples{
print(1)
}
