% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_assessment_plots.R
\name{plot_assessment_plots}
\alias{plot_assessment_plots}
\title{Plot model assessment tests}
\usage{
plot_assessment_plots(
  data,
  method = "fg",
  type = "residual_plot",
  draw_curves = "none",
  plot_theme = c("custom", "default", "noklus", "soft", "depression", "happy"),
  additional_arguments = NULL,
  testing = FALSE
)
}
\arguments{
\item{data}{\code{data table}, \code{data frame} or \code{list} - Data on long-format grouped by comparison. Must contain 'comparison', 'SampleID', 'ReplicateID', 'MP_A' and 'MP_B'.}

\item{method}{\code{character} - Method to use when calculating the linear model residuals. Can be one of the following: 'fg' (default), 'clsi' or 'ols'.}

\item{type}{\code{character} - The type of model assessment plot to be plotted. Can either be 'residual_plot' (default), 'qq_plot', 'sd_vs_concentration', 'cv_vs_concentration' or 'residual_histogram'}

\item{draw_curves}{\code{character} - Should curves be drawn to visualize relationships between two quantities (quantities may differ based on which plot is chosen by \code{type})}

\item{plot_theme}{\code{character} - Which plotting theme should be applied to the plots. Accepted inputs are 'default', 'noklus', 'soft', 'depression', and 'happy'.}

\item{additional_arguments}{\code{list} or \code{named array} with 1 row - Additional arguments to be used in calculations and plotting. Possible additional arguments are:
\itemize{
  \item{\code{x_name}:  }{Title on x-axis. Default value depends on \code{type}}
  \item{\code{y_name}:  }{Title on y-axis. Default value depends on \code{type}}
  \item{\code{n_breaks}:  }{Number of axis-ticks used. Default value is \code{6}}
  \item{\code{main_title}:  }{Plot title. Default value depends on \code{type}}
  \item{\code{sub_title}:  }{Plot subtitle. Default value is \code{NULL}}
  \item{\code{title_color}:  }{Color of plot title. Default value is \code{'black'}}
  \item{\code{sub_title_color}:  }{Color of plot subtitle. Default value is \code{'black'}}
  \item{\code{strip_fill}:  }{Fill color of the grouping labels. Default value depends on \code{type}}
  \item{\code{strip_text_color}:  }{Text color of the grouping labels. Default value is \code{'black'}}
  \item{\code{point_fill}:  }{Fill color of points. Default value is \code{'violet'}}
  \item{\code{point_border}:  }{Color of points' border. Default value is \code{'black'}}
  \item{\code{point_shape}:  }{Shape of points. Possible values are 'circle', 'square', 'diamond' and 'triangle'. Default is \code{'circle'}}
  \item{\code{histogram_fill}:  }{Fill color of histogram bins. Default value is \code{#F9F9F9} (very light gray). Only relevant if \code{type = 'residual_histogram'}}
  \item{\code{histogram_border}:  }{Border color of histogram bins. Default value is \code{'black'}. Only relevant if \code{type = 'residual_histogram'}}
  \item{\code{curve_color}:  }{Color of drawn curve. Default value is \code{'black'}}
  \item{\code{curve_se}:  }{Should confidence bands be included to drawn curve. Default value is \code{FALSE}}
  \item{\code{curve_se_fill}:  }{Color of the confidence bands included to drawn curve. Default value is \code{#F9F9F9} (very light gray)}
  \item{\code{curve_se_level}:  }{Confidence level for the confidence bands included to drawn curve. Default value is \code{0.95}}
  \item{\code{curve_formula}:  }{Edit the standard formula 'y ~ x'. Can be any 'y ~ f(x)'. Default value is 'y ~ x'}
  \item{\code{loess_span}:  }{Span of the loess curve. Default value is \code{0.90}. Larger values increase smoothing and smaller values decrease smoothing.}
  \item{\code{cv_percent}:  }{Should estimated CVS be given in percent? Default is \code{FALSE}. Only relevant if \code{type = 'cv_vs_concentration'}}
  \item{\code{var_instead}:  }{Should variance be given instead of standard deviation? Default is \code{FALSE}. Only relevant if \code{type = 'sd_vs_concentration'}}

}}

\item{testing}{\code{logical} - Should the output be test-friendly? This should only be used by the maintainer of the package, as it has no practical use to the end-user.}
}
\value{
A \code{ggplot2} object, which is the drawn assessment plot
}
\description{
Plot model assessment tests
}
\examples{
print(1)
}
