% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_commutability_evaluation_plots.R
\name{plot_commutability_evaluation_plots}
\alias{plot_commutability_evaluation_plots}
\title{Plot Commutability Evaluation Plots For Each Comparison}
\usage{
plot_commutability_evaluation_plots(
  cs_data,
  pb_data,
  ce_data,
  exclude_rings = FALSE,
  exclude_cs = FALSE,
  plot_theme = c("custom", "default", "noklus", "soft", "depression", "happy"),
  additional_arguments = NULL,
  testing = FALSE
)
}
\arguments{
\item{cs_data}{A \code{data.table}, \code{list}, or \code{data.frame}. The
mean-of-replicates (MOR) clinical sample data. Must contain:
\itemize{
  \item \code{comparison: } A \code{character} vector.
        The IVD-MD comparison identifiers. Typically on the
        form \code{'MP_A - MP_B'}.
  \item \code{SampleID: } A \code{character} vector. The
        clinical sample identifiers.
  \item \code{MP_A: } A \code{numeric} vector. The
        mean-of-replicates results from IVD-MD \code{MP_A}
        (response).
  \item \code{MP_B: } A \code{numeric} vector. The
        mean-of-replicates results from IVD-MD \code{MP_B}
        (predictor).
}}

\item{pb_data}{A \code{data.table} or \code{data.frame}. The Prediction Band
(PB) data. See \code{?do_commutability_evaluation()}.}

\item{ce_data}{A \code{data.table} or \code{data.frame}. The Commutability
Evaluation (CE) data. See
\code{?do_commutability_evaluation()}.}

\item{exclude_rings}{A \code{logical}. If \code{TRUE}, \code{inside_rate}
circles are not drawn in the plots.}

\item{exclude_cs}{A \code{logical}. If \code{TRUE}, clinical sample
mean-of-replicates results are not drawn in the plots.}

\item{plot_theme}{A \code{character} string. The desired plotting theme.
Options include:
\itemize{
   \item \code{custom: } No predefined theme. Customized
         theme may be defined by given arguments in
         \code{additional_arguments}.
   \item \code{default: } Uses default theme.
   \item \code{noklus: } Uses Noklus colors.
   \item \code{soft: } Uses pastel colors.
   \item \code{depression: } Uses dark hues such as black,
         gray, and purple.
   \item \code{happy: } Uses vibrant colors such as orange,
         red and yellow.
}}

\item{additional_arguments}{A \code{list}. Additional arguments to modify
the appearance of the output plot. The default
value is \code{NULL}, indicating that the plot
will be generated using the default settings.
Possible additional_argument entries include:
\itemize{
   \item \code{main_title: } Main title of the plot
   \item \code{sub_title: } Sub title of the plot
   \item \code{x_name: } Title of the x-axis
   \item \code{y_name: } Title of the y-axis
   \item \code{n_breaks: } Number of axis-breaks
   \item \code{title_color: } Text color of 'main_title'. Defaults to
         \code{'black'}.
   \item \code{sub_title_color: } Text color of 'sub_title'. Defaults to
         \code{'black'}
   \item \code{pb_fill:  } Fill color of prediction bands if differences in
         non-selectivity is acceptable. Defaults to \code{'green'}.
   \item \code{pb_border: } Color of the border of the prediction bands.
         Defaults to \code{'black'}.
   \item \code{curve: } Should a additional curve be drawn based on either
         the clinical sample measurements or something else. defaults to
         \code{FALSE}.
   \item \code{curve_type: } The curve type to draw. Relvant if
         \code{curve = TRUE}. Possible choices are limited to
         \code{'equivalence_curve'}, \code{'fitted_curve'} and
         \code{'flexible_curve'}.
   \item \code{curve_color: } The curve color. Relevant if
         \code{curve = TRUE}. Defaults to \code{'black'}.
   \item \code{point_shape: } The desired shape of clinical sample points.
         Possible choices include: \code{'circle'} (default), \code{'square'},
         \code{'diamond'} and \code{'triangle'}.
   \item \code{point_size: } A \code{double} larger than \code{0}. The size
         (in pts) of drawn clinical sample points. Default is selected by
         \code{ggplot()}.
   \item \code{point_fill: } The fill color of the clinical sample points.
   \item \code{point_border: } The border color of clinical sample points.
   \item \code{comparison_fill: } The fill color for the IVD-MD comparison
         strip labels.
   \item \code{comparison_text_color: } The text color for the IVD-MD
         comparison strip labels.
   \item \code{hide_prediction_intervals: } If \code{TRUE}, pointwise
         prediction intervals for evaluated material are not drawn.
   \item \code{legend_fill: } The fill color for the legend background.
         Defaults to \code{'white'}.
   \item \code{legend_text_color: } The text color for the legend font.
         Defaults to \code{'black'}.
 }}

\item{testing}{A \code{logical} value. If \code{TRUE} (not recommended), the
output will be in a form that allows testing of the output.
Primarly used for debugging and should not be used by
end-users.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Commutability Evaluation Plots For Each Comparison
}
\examples{
print(1)
}
