% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_assessment_plots.R
\name{plot_residual_hist}
\alias{plot_residual_hist}
\title{Draw Residual Histograms}
\usage{
plot_residual_hist(assessment_data, additional_arguments, draw_curves)
}
\arguments{
\item{assessment_data}{A \code{list}. Output from
\code{get_assessment_data()}.}

\item{additional_arguments}{See \code{?plot_assessment_plots()}.}

\item{draw_curves}{See \code{?plot_assessment_plots()}.}
}
\value{
A \code{ggplot2} object. The generated residual histograms.
}
\description{
This function is used to plot a residual histogram for each IVD-MD
comparison.
}
\keyword{internal}
