% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_assessment_plots.R
\name{plot_sd_vs_concentration}
\alias{plot_sd_vs_concentration}
\title{Draw Standard Deviation (or Variance) Versus Concentration Plots}
\usage{
plot_sd_vs_concentration(assessment_data, additional_arguments, draw_curves)
}
\arguments{
\item{assessment_data}{A \code{list}. Output from
\code{get_assessment_data()}.}

\item{additional_arguments}{See \code{?plot_assessment_plots()}.}

\item{draw_curves}{See \code{?plot_assessment_plots()}.}
}
\value{
A \code{ggplot2} object. The generated standard deviation (or variance)
versus concentration plots.
}
\description{
This function is used to plot a standard deviation (or variance) versus
concentration plot for each IVD-MD in \code{data}.
}
\keyword{internal}
