% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_wise_imprecision.R
\name{process_output_mwi}
\alias{process_output_mwi}
\title{Processes the Raw Output From \code{MS_wise_imprecision()}}
\usage{
process_output_mwi(
  raw_output,
  imprecision_stats_names,
  mode,
  percent,
  variance,
  rounding = 4L
)
}
\arguments{
\item{raw_output}{A object to be processed.}

\item{imprecision_stats_names}{The names found in \code{raw_output}.}

\item{mode}{A \code{character} string. The desired output structure.}

\item{percent}{A \code{logical} value. If \code{TRUE}, coefficient of
variations are presented as percentages.}

\item{variance}{A \code{logical} value. If \code{FALSE}, standard deviations
are used instead.}

\item{rounding}{An \code{integer}. The number of decimals in the output.}
}
\value{
Processed \code{raw_output}
}
\description{
This function is used to convert the raw output from
\code{MS_wise_imprecision()} to the desired output.
}
\keyword{internal}
