% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_assessment_tests.R
\name{process_output_pat}
\alias{process_output_pat}
\title{Processed Raw Output to a desired Output Structure}
\usage{
process_output_pat(
  raw_output,
  output = "visual",
  weighted = FALSE,
  rounding = 3L,
  repeat_string = "."
)
}
\arguments{
\item{raw_output}{A \code{data.table}. The temporary output of
\code{perform_assessment_tests}}

\item{output}{A \code{character} string. The desired output format.
See \code{?perform_assessment_tests}.}

\item{weighted}{A \code{logical} value. If \code{TRUE}, heteroscedasticity
test results are irrelevant, because heteroscedasticity is
addressed by the model.}

\item{rounding}{An \code{integer}. The number of digits used in the rounding
of the results.}

\item{repeat_string}{A \code{character} string. Repeated results in the
output are replaced with this string. Defaults to
\code{'.'}.}
}
\value{
A \code{data.table}, that is processed \code{perform_assessment_tests} raw
output according to the desired \code{output} format.
}
\description{
This function cleans \code{raw_output} and possibly reformats its variables
so that they are pleasing on the eyes.
}
\keyword{internal}
