% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_data.R
\name{repair_data}
\alias{repair_data}
\title{Repair a potentially broken commutability data set}
\usage{
repair_data(data, data_check, silence = 1L)
}
\arguments{
\item{data}{A \code{list}, \code{data table} or \code{data frame} - data to be attempted repaired}

\item{data_check}{\code{list} - The output from \code{check_data()}. A repair should only be attempted if \code{check_data()} says \code{data} is perfect, acceptable or questionable. For the latter alternative there is no guarantee for possible repair}

\item{silence}{\code{integer} - Should progress reports be printed. \code{1} signify no, which is the default. \code{0} signify yes.}
}
\value{
A \code{data table} that is the repaired (or unrepaired if not able to repair) data
}
\description{
Repair a potentially broken commutability data set
}
\examples{
print(1)
}
