% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_wise.R
\name{reverse_comparison_data}
\alias{reverse_comparison_data}
\title{Revert Output From \code{get_comparison_data} to Original Form}
\usage{
reverse_comparison_data(data)
}
\arguments{
\item{data}{A \code{list}, \code{data.table} or \code{data.frame}. The
IVD-MD comparison-wise data. Must contain the following
variables:
\itemize{
   \item \code{comparison: } A \code{character} vector. The
         comparison identifiers. Typically on the form
         \code{'MP_A - MP_B'}.
   \item \code{SampleID: } A \code{character} vector. The sample
         identifiers for the clinical samples.
   \item \code{ReplicateID: }A \code{character} vector. The
         replicated measurement identifiers.
   \item \code{MP_A: } A \code{numeric} vector. The observed
         measurement results from IVD-MD \code{MP_A} (response).
   \item \code{MP_B: } A \code{numeric} vector. The observed
         measurement results from IVD-MD \code{MP_A} (predictor).
}}
}
\value{
A \code{data.table}. The original data before \code{get_comparison_data} was
applied.
}
\description{
See title. Backtrack to the data before \code{get_comparison_data} is used
on it.
}
\details{
The numerical columns may not be in alphabetical order after this function
is used even if they were in alphabetical order before
\code{get_comparison_data} was applied. Use \code{repair_data()} on the
output to ensure that the newly reverted \code{data} is up to par.
}
\examples{
# Required packages
library(smooth.commutability)

# Use example data from
cs_test_data <- crp_cs_data

# Output
print(reverse_comparison_data(cs_test_data))

}
