% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_confidence_levels_lm.R
\name{simulate_confidence_levels_lm}
\alias{simulate_confidence_levels_lm}
\title{Simulate empirical confidence levels for several sets of parameters}
\usage{
simulate_confidence_levels_lm(
  N = 100L,
  n = 1,
  m = 1,
  parameters,
  level = 0.95,
  method = c("fg", "clsi", "ols"),
  attach = TRUE,
  global_mean = FALSE,
  all_inside = FALSE,
  calculate_sem = TRUE,
  percent = FALSE,
  parallel = TRUE,
  max_cores = 25
)
}
\arguments{
\item{N}{An \code{integer} larger than or equal to \code{1L} that represents the number of replicated inside checks.}

\item{n}{An \code{integer} larger than or equal to \code{1L} that represents the number of replicated confidence levels for the particular \code{parameters}.}

\item{m}{An \code{integer} larger than or equal to \code{1L}. How many external quality assessment materials should be simulated?}

\item{parameters}{A \code{data.table} with column names representing the parameters to be included in the simulations.
Must in general include \code{n}, \code{R}, \code{cvx} and \code{cvy}. If \code{type} is included, data will
be simulated from custom non-linear functions. Note that \code{cil} and \code{ciu} also must be included if \code{type} is included.
Otherwise, if \code{type} is not included, \code{simulate_eqa_data()} from the \code{fasteqa} package will be triggered, with all its choices
of parameters accepted.}

\item{level}{A \code{double} that is larger than 0, but smaller than 1. Represents the confidence level of the estimated prediction intervals for the linear regression model fit.}

\item{method}{A \code{character} vector of length 1. The method to be used for the prediction interval estimation. Can be one of the following: \code{"fg"}, \code{"clsi"}, \code{"ols"}.}

\item{attach}{A non-missing \code{logical} value (\code{TRUE} / \code{FALSE}). If set to \code{TRUE}, the relevant simulation parameters are attached to the output for reference.}

\item{global_mean}{A non-missing \code{logical} value (\code{TRUE} / \code{FALSE}). If set to \code{TRUE}, empirical confidence level is estimated taking the mean overall inside checks.}

\item{all_inside}{A non-missing \code{logical} value (\code{TRUE} / \code{FALSE}). If set to \code{TRUE} and \code{m > 1}, family-wise emprical confidence level is estimated.}

\item{calculate_sem}{A non-missing \code{logical} value (\code{TRUE} / \code{FALSE}). If set to \code{TRUE}, standard error of the empirical confidence level(s) are calculated.}

\item{percent}{A non-missing \code{logical} value (\code{TRUE} / \code{FALSE}). If set to \code{TRUE}, all numbers that are possible to present as percentages are presented as percentages.}

\item{parallel}{A non-missing \code{logical} value (\code{TRUE} / \code{FALSE}). If set to \code{TRUE}, simulations are done on multiple cores on your local machine for the duration of the function call.}

\item{max_cores}{An \code{integer} value between \code{2} and the number of cores you have on your computer. If you wish to use fewer cores than is on your computer, you should specify this here.}
}
\value{
A \code{data.table} with \code{n} rows, containing the \code{1:n} empirical confidence level estimates for the particular \code{parameters}.
}
\description{
Simulate empirical confidence levels for several sets of parameters
}
\examples{
print(1)
}
