% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_inside_lm.R
\name{simulate_inside_lm}
\alias{simulate_inside_lm}
\title{Simulate Clinical Sample Data and External Quality Assessment Data and Check Whether Inside PI}
\usage{
simulate_inside_lm(
  parameters,
  m = 1,
  level = 0.95,
  method = c("fg", "clsi", "ols")
)
}
\arguments{
\item{parameters}{A \code{list}, \code{data.table} or \code{data.frame} with column names representing the parameters to be included in the simulations.
Must in general include \code{n}, \code{R}, \code{cvx} and \code{cvy}. If \code{type} is included, data will
be simulated from custom non-linear functions. Note that \code{cil} and \code{ciu} also must be included if \code{type} is included.
Otherwise, if \code{type} is not included, \code{simulate_eqa_data()} from the \code{fasteqa} package will be triggered, with all relevant
parameters accepted.}

\item{m}{A \code{integer} larger than or equal to \code{1L}. How many external quality assessment materials should be simulated?}

\item{level}{A \code{double} that is larger than 0, but smaller than 1. Represents the confidence level of the estimated prediction intervals for the linear regression fit.}

\item{method}{A \code{character} vector of length 1. The method to be used for the prediction interval estimation. Can be one of the following: \code{"fg"}, \code{"clsi"}, \code{"ols"}.}
}
\value{
A \code{m}-dimensional \code{integer} vector with ones and zeros. \code{1} signify that an external quality assessment material is inside the estimated prediction interval, whereas \code{0} signify that it is outside.
}
\description{
Simulate Clinical Sample Data and External Quality Assessment Data and Check Whether Inside PI
}
\details{
It is important that \code{parameters} holds exactly one value for each simulation parameter. Thus, if \code{parameters} is a \code{data.table} or a \code{data.frame},
         it should only have one row. Note that this function only gives one simulation replication. To repeat the simulations multiple times, you can either use the native \code{replicate()}
         function, or something similar. Alternatively, you can use the \code{simulate_confidence_level()} function of this package to do the same.
}
\examples{
simulate_inside_lm(list(n = 20, R = 3, cvx = 0.04, cvy = 0.01, df = 5))
}
