% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_inside_lm.R
\name{simulate_insides_lm}
\alias{simulate_insides_lm}
\title{Simulate Clinical Sample Data and External Quality Assessment Data and Check Whether Inside PIs}
\usage{
simulate_insides_lm(
  N = 100L,
  m = 1,
  x = 1,
  parameters,
  level = 0.95,
  method = c("fg", "clsi", "ols"),
  attach = TRUE,
  replication_id = TRUE
)
}
\arguments{
\item{N}{An \code{integer} that represents the number of replicated inside checks.}

\item{m}{An \code{integer} larger than or equal to \code{1L}. How many external quality assessment materials should be simulated?}

\item{x}{Typically an \code{integer}, but not necessarily. Represents the identifier name for the \code{N} inside simulations.}

\item{parameters}{A \code{list}, \code{data.table} or \code{data.frame} with column names representing the parameters to be included in the simulations.
Must in general include \code{n}, \code{R}, \code{cvx} and \code{cvy}. If \code{type} is included, data will
be simulated from custom non-linear functions. Note that \code{cil} and \code{ciu} also must be included if \code{type} is included.
Otherwise, if \code{type} is not included, \code{simulate_eqa_data()} from the \code{fasteqa} package will be triggered, with all its choices
of parameters accepted.}

\item{level}{A \code{double} that is larger than 0, but smaller than 1. Represents the confidence level of the estimated prediction intervals for the linear regression model fit.}

\item{method}{A \code{character} vector of length 1. The method to be used for the prediction interval estimation. Can be one of the following: \code{"fg"}, \code{"clsi"}, \code{"ols"}.}

\item{attach}{A non-missing \code{logical} value (\code{TRUE} / \code{FALSE}). If set to \code{TRUE}, the relevant simulation parameters are attached to the output for reference.}

\item{replication_id}{A non-missing \code{logical} value (\code{TRUE} / \code{FALSE}). If set to \code{TRUE}, inside check replications for each \code{1:N} is given an identifier in the output. Particularly relevant if \code{m > 1}.}
}
\value{
A \code{data.table} containing the inside replications for the inputs.
}
\description{
Simulate Clinical Sample Data and External Quality Assessment Data and Check Whether Inside PIs
}
\examples{
simulate_insides(parameters = list(n = 25, R = 3, cvx = 0.01,
 cvy = 0.04, qran = 0.3, qpos = 1, mmax = 5, df = 5))
}
