% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_zetas.R
\name{simulate_zetas}
\alias{simulate_zetas}
\title{Monte Carlo Simulations for the Estimator of zeta}
\usage{
simulate_zetas(
  n,
  parameters,
  statistic = median,
  m = 1L,
  attach = TRUE,
  probs = 0.99,
  simplify = FALSE
)
}
\arguments{
\item{n}{An \code{integer} equal to or larger than 1, specifying the number of zeta value simulations for each replication of the target \code{statistic}.}

\item{parameters}{A \code{list}, \code{data.table}, or \code{data.frame} with simulation parameters. Refer to \code{?simulate_eqa_data} for detailed options and descriptions.}

\item{statistic}{An \code{R function} applied to the simulated zeta values or a \code{character} string selecting from predefined options to approximate various statistics of the estimated zeta value. Valid character strings are:
\itemize{
  \item \code{none}: Outputs the raw simulated values.
  \item \code{all}: Computes the first four moments and the following percentiles: 1st, 2.5th, 5th, 10th, 25th, 50th, 75th, 90th, 95th, 97.5th, and 99th.
  \item \code{all2}: Computes the first four moments and the 1st and 99th percentiles.
  \item \code{moments}: Computes the first four moments.
  \item \code{moments12}: Computes the first two moments.
  \item \code{moments34}: Computes the third and fourth moments.
  \item \code{quantiles}: Computes the following percentiles: 1st, 2.5th, 5th, 10th, 25th, 50th, 75th, 90th, 95th, 97.5th, and 99th.
  \item \code{quantiles2}: Computes the following percentiles: 1st, 25th, 50th, 75th, and 99th.
  \item \code{quartiles}: Computes the 25th, 50th, and 75th percentiles (i.e., the quartiles).
}}

\item{m}{An \code{integer} ≥ 1, indicating the number of replications for the \code{statistic}. If \code{m > 1}, \code{n} zeta values are simulated \code{m} times. For each set of \code{n} simulated values, the \code{statistic} is computed, yielding \code{m} replications. Be cautious: large \code{n} and \code{m} values significantly impact computation time. It is advised not to let n times m exceed \code{1e7L} to prevent excessive runtimes.}

\item{attach}{A non-missing \code{logical} value. If \code{TRUE} (default), the computed statistic(s) for the zeta value are added to the \code{parameters} object for easy reference.}

\item{probs}{A \code{double} between 0 and 1. Relevant when \code{statistic} is \code{quantile}. For example, to simulate the 99th percentile of zeta's estimator, set \code{statistic = "quantile"} and \code{probs = 0.99}.}

\item{simplify}{A non-missing \code{logical} value. If \code{TRUE}, the output will be attempted to be simplified. If \code{attach} is \code{TRUE} and \code{simplify} is \code{TRUE} the output will be always be simplified to a \code{data.table} object. On the other hand, if \code{attach} is \code{FALSE} the output will only be simplified to a \code{data.table} if several statistics are simulated, and simplified to a vector if only one statistic is simulated.}
}
\value{
A \code{list} either with both the input parameters and simulated statistics for zeta's estimator or just the simulated statistics, based on the \code{attach} option.
}
\description{
Simulate statistics of the zeta estimator using Monte Carlo methods.
}
\examples{
# Simulate 100 standard deviations of zeta's estimator with 'eta' = 5 (heteroscedasticity).
# Simulations use a design with n = 20 clinical samples and R = 4 replicates.
simulate_zetas(n = 100, parameters = list(n = 20, R = 4, eta = 5, eta0 = 1), statistic = sd, attach = TRUE)

# Simulate 1000 99th percentiles of zeta's estimator with 'prop' = 0.15 and 'mmax' = 10 (random DINS).
# Simulations use a design with n = 25 clinical samples and R = 3 replicates.
simulate_zetas(n = 1000, parameters = list(n = 25, R = 3, prop = 0.15, mmax = 10), statistic = quantile, attach = TRUE, probs = 0.99)
}
