% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_comparison_data}
\alias{transform_comparison_data}
\title{Apply a Special Transformation to One IVD-MD comparison}
\usage{
transform_comparison_data(data, transformation)
}
\arguments{
\item{data}{A \code{data.table}. The IVD-MD comparison results for a
particular IVD-MD comparison.}
}
\value{
A \code{data.table}. The transformed data for a particular IVD-MD
comparison.
}
\description{
This function is used to apply one of the special transformations to each
IVD-MD comparison separately. Used internally in \code{transform_data}
if one of the special transformations are desired.
Not to be used by end-users.
}
\keyword{internal}
