% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typo_suggestions.R
\name{typo_suggestions}
\alias{typo_suggestions}
\title{Suggestions for typos}
\usage{
typo_suggestions()
}
\value{
A list containing the word book for relevant words
}
\description{
Suggestions for typos
}
\examples{
typo_suggestions()
}
