% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_commutability_evaluation_plots.R
\name{update_all_attributes}
\alias{update_all_attributes}
\title{Update Plotting Attributes: PCEP}
\usage{
update_all_attributes(additional_arguments)
}
\arguments{
\item{additional_arguments}{A \code{list}.}
}
\value{
A \code{list}. The modified \code{additional_arguments}.
}
\description{
Update Plotting Attributes: PCEP
}
\keyword{internal}
