% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_assessment_plots.R
\name{update_all_attributes_pap}
\alias{update_all_attributes_pap}
\title{Applies Custom Settings and Merge Them with Defaults of \code{additional_arguments}}
\usage{
update_all_attributes_pap(additional_arguments, method, type)
}
\arguments{
\item{additional_arguments}{See \code{?plot_assessment_plots()}.}

\item{method}{See \code{?plot_assessment_plots()}.}

\item{type}{See \code{?plot_assessment_plots()}.}
}
\value{
The updated \code{additional_arguments} ready to be used.
}
\description{
This function is used to update the \code{additional_arguments} argument
based on custom choices and default choices in
\code{plot_assessment_plots()}.
}
\keyword{internal}
