% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_imprecision_data.R
\name{validate_B_eid}
\alias{validate_B_eid}
\title{Validates \code{B} parameter}
\usage{
validate_B_eid(B, invalid_NA = FALSE)
}
\arguments{
\item{B}{A numeric value representing bootstrap iterations.}

\item{invalid_NA}{A \code{logical} value. If \code{TRUE}, \code{NA} is
returned instead of throwing an error. Default is \code{FALSE}.}
}
\value{
A \code{list}. The processed \code{B} and which action should be taken
based on the value of \code{B}.
}
\description{
This function validates the \code{B} parameter (bootstrap iterations)
in \code{estimate_imprecision_data}.
}
\keyword{internal}
