% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_assessment_tests.R
\name{validate_B_pat}
\alias{validate_B_pat}
\title{Validate and Convert \code{B}}
\usage{
validate_B_pat(B)
}
\arguments{
\item{B}{An \code{integer}. See \code{?perform_assessment_tests()}.}
}
\value{
The validated and possibly coerced \code{B}
}
\description{
This function is used to validate and coerce the \code{B} argument
in \code{perform_assessment_tests}.
}
\keyword{internal}
