% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_data}
\alias{validate_data}
\title{Validate data}
\usage{
validate_data(data)
}
\arguments{
\item{data}{\code{list} or \code{data table} that is the data that we wish to validate}
}
\value{
A \code{list} of validation results
}
\description{
Validate data
}
\examples{
print(1)
}
