% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_commutability_evaluation.R
\name{validate_data_dce}
\alias{validate_data_dce}
\title{Validates \code{data} and \code{new_data}}
\usage{
validate_data_dce(data, na_rm = TRUE)
}
\arguments{
\item{data}{A object to be validated.}

\item{na_rm}{A \code{logical} value. If \code{TRUE}, \code{NA} values are
removed from the validated \code{data}.}
}
\value{
Processed \code{data} or \code{new_data}
}
\description{
This function is used to validate and convert the \code{data} or
\code{new_data} arguments in \code{do_commutability_evaluation}.
}
\keyword{internal}
