% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_imprecision_data.R
\name{validate_data_eid}
\alias{validate_data_eid}
\title{Validates \code{data}}
\usage{
validate_data_eid(data)
}
\arguments{
\item{data}{A object to be validated.}
}
\value{
Processed \code{data}
}
\description{
This function is used to validate and convert the \code{data} argument
in \code{estimate_imprecision_data}.
}
\keyword{internal}
