% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prediction_data.R
\name{validate_data_epd}
\alias{validate_data_epd}
\title{Validates \code{data}}
\usage{
validate_data_epd(data, na_rm = TRUE)
}
\arguments{
\item{data}{A object to be validated.}

\item{na_rm}{A \code{logical} value. If \code{TRUE}, \code{NA} values are
removed from the validated \code{data}.}
}
\value{
Processed \code{data}
}
\description{
This function is used to validate and convert the \code{data} argument
in \code{estimate_prediction_data}.
}
\keyword{internal}
