% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_assessment_tests.R
\name{validate_data_pat}
\alias{validate_data_pat}
\title{Validate and Converts \code{data}}
\usage{
validate_data_pat(data)
}
\arguments{
\item{data}{A object to be validated and converted.}
}
\value{
The converted \code{data}
}
\description{
This function is used to validate and convert the \code{data} argument
in \code{perform_assessment_tests}.
}
\keyword{internal}
