% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{validate_date_td}
\alias{validate_date_td}
\title{Validates and Converts The \code{data} Argument}
\usage{
validate_date_td(data, special_transformation = FALSE)
}
\arguments{
\item{data}{An object to be validated and converted to \code{data.table}.}

\item{special_transformation}{A \code{logical} value. If \code{TRUE},
\code{data} is checked according to
requirements necessary for one of the
special transformations.}
}
\value{
A \code{data.table}. Processed and validated \code{data}.
}
\description{
This function is used to validate and convert the \code{data} argument
in \code{transform_data}. Not to be used by end-users.
}
\keyword{internal}
