% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_wise_imprecision.R
\name{validate_imprecision_data_mwi}
\alias{validate_imprecision_data_mwi}
\title{Validates \code{imprecision_data}}
\usage{
validate_imprecision_data_mwi(imprecision_data)
}
\arguments{
\item{imprecision_data}{A object to be validated.}
}
\value{
Processed \code{imprecision_data}
}
\description{
This function is used to validate and convert the \code{imprecision_data}
argument in \code{MS_wise_imprecision}.
}
\keyword{internal}
