% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_assessment_tests.R
\name{validate_level_pat}
\alias{validate_level_pat}
\title{Validate and Convert \code{level}}
\usage{
validate_level_pat(level)
}
\arguments{
\item{level}{A \code{double}. See \code{?perform_assessment_tests()}.}
}
\value{
The validated and possibly coerced \code{level}
}
\description{
This function is used to validate and coerce the \code{level} argument
in \code{perform_assessment_tests}.
}
\keyword{internal}
