% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_assessment_plots.R
\name{validate_method}
\alias{validate_method}
\title{Validates and Converts The \code{method} Argument}
\usage{
validate_method(method)
}
\arguments{
\item{method}{A \code{character} string to be validated and converted.}
}
\value{
The converted \code{data}
}
\description{
This function is used to validate and convert the \code{method} argument
in \code{plot_assessment_plots()}.
}
\keyword{internal}
