% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prediction_data.R
\name{validate_new_data_epd}
\alias{validate_new_data_epd}
\title{Validates \code{new_data}}
\usage{
validate_new_data_epd(new_data, na_rm = TRUE)
}
\arguments{
\item{new_data}{A object to be validated.}

\item{na_rm}{A \code{logical} value. If \code{TRUE}, \code{NA} values are
removed from the validated \code{new_data}.}
}
\value{
Processed \code{new_data}
}
\description{
This function is used to validate and convert the \code{new_data} argument
in \code{estimate_prediction_data}.
}
\keyword{internal}
