% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_assessment_plots.R
\name{validate_plot_theme}
\alias{validate_plot_theme}
\title{Validates and Converts The \code{plot_theme} Argument}
\usage{
validate_plot_theme(plot_theme)
}
\arguments{
\item{plot_theme}{A \code{character} string to be validated and converted.}
}
\value{
The converted \code{data}
}
\description{
This function is used to validate and convert the \code{plot_theme} argument
in \code{plot_assessment_plots()}.
}
\keyword{internal}
