% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_assessment_plots.R
\name{validate_type}
\alias{validate_type}
\title{Validates and Converts The \code{type} Argument}
\usage{
validate_type(type)
}
\arguments{
\item{type}{A \code{character} string to be validated and converted.}
}
\value{
The converted \code{data}
}
\description{
This function is used to validate and convert the \code{type} argument
in \code{plot_assessment_plots()}.
}
\keyword{internal}
