% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check if Data is Suitable for Further Use in This Package}
\usage{
check_data(data, type = "cs")
}
\arguments{
\item{data}{A \code{data.table}, \code{list} or \code{data.frame}. The
dataset to be validated.}

\item{type}{A \code{character} string. The type of \code{data} must be
one of the following:
\itemize{
   \item \code{'cs': } \code{data} is treated as clinical sample
         data.
   \item \code{'eqam': } \code{data} is treated as evaluated
         material data.
}
The suitability requirements for \code{data} will depend on
\code{type}. For example, there are different requirements for
clinical sample data and external quality assessment material
data.}
}
\value{
A \code{list} of length \code{3}.
}
\description{
Check if Data is Suitable for Further Use in This Package
}
\examples{
print(1)
}
