% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_commutability_evaluation.R
\name{do_commutability_evaluation}
\alias{do_commutability_evaluation}
\title{Run a complete commutability evaluation based on CS data and EQA/CR material data}
\usage{
do_commutability_evaluation(
  data,
  new_data,
  method_pi = "fg",
  method_bs = "percentile",
  level_pi = 0.99,
  level_bs = 0.95,
  M = 0,
  upper_zeta = 2.25,
  output = "sufficient"
)
}
\arguments{
\item{data}{\code{list}, \code{data table} or \code{data frame} - Clinical sample data (CS data) for the commutability assessment experiment}

\item{new_data}{\code{list}, \code{data table} or \code{data frame} - Evaluated samples' data (EQAM / CRM data) for the commutability assessment experiment}

\item{method_pi}{\code{character} - Which method should be used to estimate the Deming prediction intervals. Default is \code{fg} which is Fuller and Gillard's approach. Alternatively, \code{clsi} may be used for EP14's approach}

\item{method_bs}{\code{character} - Which bootstrap method should be used to estimated the confidence levels of zeta and imprecision estimates}

\item{level_pi}{\code{double} - Number between 0 and 1, which is the required confidence level for the estimated prediction intervals}

\item{level_bs}{\code{double} - Number between 0 and 1, which is the required confidence level for the estimated bootstrap confidence intervals}

\item{M}{\code{double} - If upper_zeta is not given, how much differences in non-selectivity between IVD-MD pairs should we accept in terms of relative increase in prediction interval width. For example, 0.15 signify that we accept an average of 15 \% relative increase}

\item{upper_zeta}{\code{double} - Either NULL or a number. A estimated value of zeta above \code{upper_zeta} results in concluding with excessive differences in non-selectivity between IVD-MD pairs. If \code{NULL}, M will be used to look for an appropriate upper limit utilizing Monte Carlo simulations}

\item{output}{\code{character} - Either \code{sufficient} or \code{complete}. Passing \code{sufficient} will only output the necessary data analysis for the commutability evaluation analysis. Passing \code{complete} will output all data analysis for the commutability evaluation analysis}
}
\value{
A two-dimensional \code{list} containing \code{merged_pb_data} and \code{merged_ce_data}, and these are of \code{data table} class. \code{merged_pb_data} is used for plotting, whereas \code{merged_ce_data} is the summary of the data analysis concerning the commutability evaluation experiment
}
\description{
Run a complete commutability evaluation based on CS data and EQA/CR material data
}
\examples{
print(1)
}
