% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_outlier_analysis.R
\name{do_outlier_analysis}
\alias{do_outlier_analysis}
\title{Outlier Analysis Based on Clinical Sample Data}
\usage{
do_outlier_analysis(
  data,
  method = "qrange",
  variable = "influence",
  level = 0.99,
  output = "visual"
)
}
\arguments{
\item{data}{A \code{data.table}, \code{list} or \code{data frame}. The
clinical sample measurement data. Must contain the following
variables:
\itemize{
   \item \code{comparison: } A \code{character} vector. The
         comparison identifiers. Typically on the form
         \code{'MP_A - MP_B'}.
   \item \code{SampleID: }  A \code{character} vector. The sample
         identifiers for the clinical samples.
   \item \code{ReplicateID: } A \code{character} vector. The
         replicated measurement identifiers.
   \item \code{MP_A: } A \code{numeric} vector. The observed
         measurement results from IVD-MD \code{MP_A} (response).
   \item \code{MP_B: } A \code{numeric} vector. The observed
         measurement results from IVD-MD \code{MP_A} (predictor).
}}

\item{method}{A \code{character} string. The desired outlier detection method
 to apply to \code{data}. Options include:
 \itemize{
    \item \code{qrange: } Studentized Q-range test. Will be
          performed on \code{ReplicateID}-level.
    \item \code{burnett: } Burnetts method. Default. Will be
          performed on \code{SampleID}-level.
    \item \code{chauvenet: } Chauvenet's Criterion. Will be
          performed on \code{SampleID}-level. Only statistically
          significant outliers according to the Tietjen-Moore
          test will be considered outliers among the
          \code{chauvenet} candidates.
    \item \code{iqr: } Inter-Quartile Range test. Will be
          performed on \code{SampleID}-level. Only statistically
          significant outliers according to the Tietjen-Moore
          test will be considered outliers among the
          \code{iqr} candidates.
}}

\item{variable}{A \code{character} string. The variable to be assessed when
looking for outliers. Possible choices are:
\itemize{
   \item \code{influence: } Uses studentized residuals
         from a particular non-linear model. Default.
   \item \code{raw_difference: } Uses the raw differences
         between \code{MP_A} and \code{MP_B}.
         That is, \eqn{d_i = y_i - x_i}. Not recommended.
   \item \code{absolute_difference: } Uses the absolute
         difference between \code{MP_A} and \code{MP_B}.
         That is, \eqn{d_i = |y_i - x_i|}. Not recommended.
   \item \code{relative_difference: } Uses the relative
         difference between \code{MP_A} and \code{MP_B}
         measurements. That is,
         \eqn{d_i = \frac{|y_i - x_i|}{|x_i|} \cdot 100\%}.
   \item \code{log_difference: } Uses the log difference
         between \code{MP_A} and \code{MP_B}. That is,
         \eqn{d_i = \log(y_i) - \log(x_i)}.
}
Ignored if \code{method = 'qrange'}.}

\item{level}{A \code{double}. Must be between \code{0} and \code{1}. The
desired confidence level. Defaults to \code{0.99}
(\eqn{99\%}).}

\item{output}{A \code{character} string. The output type.}
}
\value{
a \code{data.table}. The dimensions depend on \code{method}, \code{variable}
and \code{output}.
}
\description{
Performs outlier tests that is meant to check for multiple outliers at once.
}
\details{
For IQR (\code{method = 'iqr'}) and Chauvenet (\code{method = 'chauvenet'})
outlier detection methods, candidate outliers are further assessed using the
Tietjen-Moore test at a significance level of \code{1 - level}.
This additional check helps mitigate the sensitivity of these methods, which
can sometimes flag too many data points as outliers.

Burnett's method (\code{method = 'burnett'}) does not typically suffer from
the same sensitivity issue, so no additional tests are applied.

Important Note: Outliers identified by any of these methods should not
be automatically removed without further investigation. In some cases,
flagged data points may represent meaningful variations in the underlying
distribution. Careful consideration should be given before deciding to
exclude them.

The format of the function output is controlled by the \code{output}
argument:
\itemize{
   \item \code{'visual': } Use to presen results concisely with a focus on
         interpretability.
   \item \code{'raw': } Use when further calculations or custom processing
         of the detected outliers is required.
}
}
\examples{
# Required packages
library(smooth.commutability)
library(data.table)

# Use example data from smooth.commutability package
test_cs_data <- copy(crp_cs_data)

# Create an extreme value
test_cs_data$MP_A[1:3] <- test_cs_data$MP_A[1:3] + 26

# Outlier test using standardized residuals and Burnett's method
do_outlier_analysis(
 data = test_cs_data,
 method = "burnett",
 variable = "influence",
 level = 0.95,
 output = "visual"
)

}
