% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_imprecision_data.R
\name{estimate_imprecision_data}
\alias{estimate_imprecision_data}
\title{All calculations regarding imprecision}
\usage{
estimate_imprecision_data(
  data,
  B = 2000L,
  type = "percentile",
  level = 0.95,
  invalid_NA = FALSE
)
}
\arguments{
\item{data}{A \code{list}, \code{data.table} or \code{data.frame}. The
CS data. Must contain the following variables:
\itemize{
   \item \code{comparison: } A \code{character} vector. The
         comparison identifiers. Typically on the form
         \code{'MP_A - MP_B'}.
   \item \code{SampleID: } A \code{character} vector. The sample
         identifiers for the clinical samples.
   \item \code{ReplicateID: }A \code{character} vector. The
         replicated measurement identifiers.
   \item \code{MP_A: } A \code{numeric} vector. The observed
         measurement results from IVD-MD \code{MP_A} (response).
   \item \code{MP_B: } A \code{numeric} vector. The observed
         measurement results from IVD-MD \code{MP_A} (predictor).
}}

\item{B}{An \code{integer}. Must be larger than or equal to \code{50}. The
Number of bootstrap replicates used to estimate bootstrap
confidence intervals for each of the repeatability imprecision
statistics. Defaults to \code{2e3L} (2,000). If \code{NULL},
estimated confidence intervals are not calculated.}

\item{type}{A \code{character} string. Determines the bootstrap method to
 apply for estimating the confidence intervals:
 \itemize{
    \item \code{normal: } Standard normal bootstrap confidence
                          interval.
    \item \code{basic: } Basic bootstrap confidence interval.
    \item \code{percentile: } Percentile bootstrap confidence
                              interval.
    \item \code{BCa: } Bias- and skewness-corrected bootstrap
                       confidence interval.
}}

\item{level}{A \code{double}. Must be between \code{0} and \code{1}. The
desired nominal confidence level for the bootstrap estimated
confidence intervals. The default is \code{0.95} (\eqn{95\%}).}

\item{invalid_NA}{A \code{logical} value. If \code{TRUE} (not recommended),
return \code{NA} rather than throwing an error if
something goes wrong.}
}
\value{
A \code{data.table} with \code{unique(data$comparison)} rows and \code{5} or
\code{15} columns. The columns are named: \code{CV_A}, \code{CV_A_lwr},
\code{CV_A_upr}, \code{CV_B}, \code{CV_B_lwr}, \code{CV_B_upr},
\code{lambda}, \code{lambda_lwr}, \code{lambda_upr}, \code{Var_A},
\code{Var_A_lwr}, \code{Var_A_upr}, \code{Var_B}, \code{Var_B_lwr},
\code{Var_B_upr}.
}
\description{
Obtains all necessary information on repeatability imprecision for each
unique IVD-MD pair (\code{comparison}) in \code{data}.
}
\details{
Estimate \code{Var_A}, \code{Var_B}, \code{CV_A}, \code{CV_B} and
\code{lambda}, with corresponding bootstrap confidence intervals for each
unique \code{comparison} element in \code{data}.

This function is generally not recommended to be used by end-users. It is
however an extremely important helping function that is used in the
\code{do_commutability_evaluation()} wrapper function.
}
\examples{
print(1)
}
