% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_imprecision_data.R
\name{estimate_imprecision_data}
\alias{estimate_imprecision_data}
\title{All calculations regarding imprecision}
\usage{
estimate_imprecision_data(data, B = 2000, type = "percentile", level = 0.95)
}
\arguments{
\item{data}{\code{list}, \code{data frame} or \code{data table} that is grouped by \code{"comparison"}. Furthermore, must contain the following columns: \code{comparison}, \code{SampleID}, \code{ReplicateID}, \code{MP_A} and \code{MP_B}}

\item{B}{Number of bootstrap replicates used to estimate bootstrap confidence intervals. The default is 2000, which is the typical for bootstrap confidence intervals. Note that if you have more than five unique IVD-MDs in \code{data}, it may take approximately two seconds to run the resampling}

\item{type}{Type of bootstrap confidence interval. There are four options:
\itemize{
  \item{\code{normal}: }{Standard normal bootstrap confidence interval}
  \item{\code{basic}: }{Basic bootstrap confidence interval}
  \item{\code{percentile}: }{Percentile bootstrap confidence interval}
  \item{\code{BCa}: }{Not yet available}
}}

\item{level}{Confidence level of bootstrap confidence interval. A 95 percent confidence level is the default}
}
\value{
A \code{data table} with entries \code{CV_A}, \code{CV_A_lwr}, \code{CV_A_upr}, \code{CV_B}, \code{CV_B_lwr}, \code{CV_B_upr}, \code{lambda}, \code{lambda_lwr}, \code{lambda_upr}, \code{Var_A}, \code{Var_A_lwr}, \code{Var_A_upr}, \code{Var_B}, \code{Var_B_lwr}, \code{Var_B_upr}
}
\description{
Obtain all necessary information on imprecision data for each unique pair of IVD-MDs in your data. The output is on the form required by \code{merge_results}(), making it very useful.
}
\examples{
print(1)
}
