% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prediction_data.R
\name{estimate_prediction_data}
\alias{estimate_prediction_data}
\title{Estimate prediction intervals for Deming regression based on EQA data}
\usage{
estimate_prediction_data(
  data,
  new_data = NULL,
  method = "fg",
  level = 0.99,
  rounding = 3L
)
}
\arguments{
\item{data}{\code{list}, \code{data table} or \code{data frame} - Data containing clinical samples' measurements}

\item{new_data}{\code{list}, \code{data table} or \code{data frame} - Data containing either external quality assessment materials' measurements or new values for \code{MP_B}. Including \code{SampleID}, \code{MP_A}, \code{MP_B} will create commutability evaluation data based on control material measurements, whereas having only \code{MP_B} yields prediction band data}

\item{method}{\code{character} - 1 x 1 character signifying which method should be applied for estimating the prediction intervals}

\item{level}{\code{float} - Number between 0 and 1, that is the confidence level of the estimated prediction intervals}

\item{rounding}{\code{integer} - How many decimals should be returned in the output. Default is \code{3}}
}
\value{
\code{data table} containing commutability evaluation data or prediction band data depending on the given new_data
}
\description{
Estimate prediction intervals for Deming regression based on EQA data
}
\examples{
print(1)
}
