% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_look_up_table.R
\docType{data}
\name{look_up_table}
\alias{look_up_table}
\title{Look-up table for upper \eqn{\hat{\zeta}} values}
\format{
A data.table with 2604 rows and 4 variables:
\describe{
  \item{n}{The number of clinical samples in the study design}
  \item{R}{The number of replicated measurements performed on each clinical sample in the study design}
  \item{M}{Average relative increase of point-wise prediction intervals attributed to differences in non-selectivity}
  \item{zeta}{The associated lower limit of the rejection region of \eqn{\hat{\zeta}} for n, R and M}
}
}
\usage{
look_up_table
}
\description{
A dataset containing critical values of the differences in non-selectivity estimator, \eqn{\hat{\zeta}}, based on 2604 unqiue combinations of study designs and
maximum tolerable differences in non-selectivity magnitudes. That is, all combinations of \eqn{n = 20, 21, \ldots, 50}, \eqn{R = 2, 3, 4, 5}, and \eqn{M = 0, 0.05, 0.10, \ldots, 0.95, 1}.
}
\keyword{datasets}
