% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_wise.R
\name{MS_wise}
\alias{MS_wise}
\title{MS_wise}
\usage{
MS_wise(data)
}
\arguments{
\item{data}{\code{list} or \code{data table} containing \code{SampleID}, \code{ReplicateID}, and measurement result columns}
}
\value{
A \code{data table} containing \code{data} on long format
}
\description{
MS_wise
}
\examples{
print(1)
}
