% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_wise_imprecision.R
\name{MS_wise_imprecision}
\alias{MS_wise_imprecision}
\title{Get MS-wise imprecison estimates}
\usage{
MS_wise_imprecision(imprecision_data, mode = c("visual", "exact"))
}
\arguments{
\item{imprecision_data}{\code{list}, \code{data table} or \code{data frame} - Imprecision estimates including bootstrap confidence intervals for CV data, variance data, or both. Must include a grouping column named \code{comparison}.}

\item{mode}{character - How should the output be expressed. There are two possibilities:
\itemize{
  \item{\code{visual}: }{Creates presentable MS-wise imprecision estimates, that neatly combines MS-wise imprecision estimates into a maximum of three columns. The first column is of course \code{MS}, and the remaining potential two is all imprecision estimates for CVs and SDs}
  \item{\code{exact}: }{Creates MS-wise imprecision estimates that could be used for further calculations if needed. Includes a minimum of four columns (and a maxium of 7), where the latter six rows are the imprecision estimates for each MS}
}}
}
\value{
\code{data table} with MS-wise imprecision estimates, that include point estimates and bootstrap confidence intervals
}
\description{
Get MS-wise imprecison estimates
}
\examples{
print(1)
}
