% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_assessment_tests.R
\name{perform_assessment_tests}
\alias{perform_assessment_tests}
\title{Perform linear model assessment tests}
\usage{
perform_assessment_tests(
  data,
  B = 2000L,
  method = "fg",
  level = 0.95,
  include_rejection_rates = TRUE,
  simultaneous_testing = TRUE,
  silence = 1L
)
}
\arguments{
\item{data}{\code{data table}, \code{list} or \code{data frame} - Clinical sample data grouped by 'comparison'. This data must include all replicated measurements, or else imprecision estimates may not be calculated.}

\item{B}{\code{integer} - The number of bootstrap replicates used to calculate the rejection rates of the assessment tests}

\item{method}{\code{character} - Which method should be used to calculate the linear model residuals used in the normality tests. Possible inputs are \code{'fg'} (default), \code{'clsi'} and \code{'ols'}. \code{'fg'} uses Deming regression defined by Fuller & Gillard. \code{'clsi'} uses Deming regression defined by CLSI EP14. \code{'ols'} uses ordinary least squared linear regression.}

\item{level}{\code{double} - A value between 0 and 1, that corresponds with the significance level of the assessment tests. Default level is \code{0.95}.}

\item{include_rejection_rates}{\code{logical} - \code{TRUE} or \code{FALSE}, where the former is the default. Setting this to \code{TRUE} ensures that assessment test rejection rates are included in the output. Setting this to \code{FALSE}, disables any bootstrap estimation and exclude rejection rates from the output.}

\item{simultaneous_testing}{\code{logical} - \code{TRUE} or \code{FALSE}, where the former is the default. Setting this to \code{TRUE}, entails that correction for multiple testing is performed. On the contratry, setting this to \code{FALSE}, will use 1 - level for all assessment tests.}

\item{silence}{\code{integer} - Should verbose be allowed? if silence is larger or equal to \code{1}, verbose is disabled. If silence is smaller than \code{1}, verbose is enabled.}
}
\value{
A \code{data table} with results concerning the performed assessment tests
}
\description{
Perform linear model assessment tests
}
\examples{
print(1)
}
