% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_assessment_plots.R
\name{plot_assessment_plots}
\alias{plot_assessment_plots}
\title{Plot Model Diagnostics Plots}
\usage{
plot_assessment_plots(
  data,
  method = "fg",
  type = "residual_plot",
  draw_curves = TRUE,
  plot_theme = "default",
  additional_arguments = NULL
)
}
\arguments{
\item{data}{A \code{data.table}, \code{list} or \code{data.frame}.
Must contain the clinical sample data with variables:
\itemize{
   \item \code{comparison: } A \code{character} vector. The
         comparison identifiers. Typically on the form
         \code{'MP_A - MP_B'}.
   \item \code{SampleID: } A \code{character} vector. The sample
         identifiers for the clinical samples.
   \item \code{ReplicateID: } A \code{character} vector. The
         replicated measurement identifiers.
   \item \code{MP_A: } A \code{numeric} vector. The observed
         measurement results from IVD-MD \code{MP_A} (response).
   \item \code{MP_B: } A \code{numeric} vector. The observed
         measurement results from IVD-MD \code{MP_A} (predictor).
}}

\item{method}{A \code{character} string. The desired model for calculating
residuals and fitted values. The possible options are:
\code{'fg'} (default), \code{'clsi'}, \code{'ols'},
\code{'ss'} and \code{'ssw'}.}

\item{type}{A \code{character} string. The desired model assement plot to
draw. Can be one out of:
\itemize{
 \item \code{residual_plot: } Residual plot (default).
       Standardized model residuals at the y-axis and fitted
       values at the x-axis.
 \item \code{residual_histogram: } Residual histogram plot.
       Standardized residuals are the x-axis and density values
       at the y-axis.
 \item \code{qq_plot: } Normal QQ plot. Theoretical quantiles
       of the N(0, 1)-distribution at the x-axis and empirical
       quantiles of the standardized residuals at the y-axis.
 \item \code{sd_vs_concentration: } Clinical sample standard
       deviations at the y-axis and the corresponding
       concentration values at the x-axis.
 \item \code{cv_vs_concentration: } Clinical sample coefficient
       of variation at the y-axis and the corresponding
       concentration values at the y-axis.
}}

\item{draw_curves}{A \code{logical} value. If \code{TRUE} (default),
informative curves are added to the assessment plots when
relevant.}

\item{plot_theme}{A \code{character} string. The desired plotting theme.
Options include:
\itemize{
   \item \code{custom: } No predefined theme. Customized
         theme may be defined by given arguments in
         \code{additional_arguments}.
   \item \code{default: } Uses default theme.
   \item \code{noklus: } Uses Noklus colors.
   \item \code{soft: } Uses pastel colors.
   \item \code{depression: } Uses dark hues such as black,
         gray, and purple.
   \item \code{happy: } Uses vibrant colors such as orange,
         red and yellow.
}}

\item{additional_arguments}{A \code{list}. Additional arguments to modify
the appearance of the output plot. The default
value is \code{NULL}, indicating that the plot
will be generated using the default settings.
Possible additional_argument entries include:
\itemize{
 \item \code{main_title: }
 \item \code{sub_title: }
 \item \code{x_name: }
 \item \code{y_name: }
 \item \code{n_breaks: }
 \item \code{title_color: }
 \item \code{sub_title_color: }
 \item \code{strip_fill: }
 \item \code{strip_text_color: }
 \item \code{point_shape: }
 \item \code{point_size: }
 \item \code{point_fill: }
 \item \code{point_border: }
 \item \code{curve_color: }
 \item \code{curve_se: }
 \item \code{curve_se_fill: }
 \item \code{curve_formula: }
 \item \code{loess_span: }
 \item \code{histogram_fill: }
 \item \code{histogram_border: }
 \item \code{cv_percent: }
 \item \code{var_instead: }
}}
}
\value{
A \code{ggplot2} object. The drawn desired assessment plots.
}
\description{
Generate model assessment plots for each IVD-MD comparison (or IVD-MD).
}
\details{
Modeling make assumptions. One should strive to test these assumptions to
validate whether a particular estimated model can be trusted. This function
can be used to draw different diagnostic plots that be used to assess
whether the model assumptions are satisfied for \code{data}.

Currently, this function supports five different plot types:
\itemize{
   \item \code{residual_plot: } Residual plot (default). Standardized model
         residuals at the y-axis and fitted values at the x-axis. Can be
         used to check for model error heteroscedasticity. Can also be used
         to check for non-linearity if a linear model is used to estimate
         the residuals.
   \item \code{residual_histogram: } Residual histogram plot. Standardized
         residuals are the x-axis and density values at the y-axis. Can be
         used to check for model error normality.
   \item \code{qq_plot: } Normal QQ plot. Theoretical quantiles of the
         N(0, 1)-distribution at the x-axis and empirical quantiles of the
         standardized residuals at the y-axis. Can be used to check for
         model error normality.
   \item \code{sd_vs_concentration: } Clinical sample standard deviations at
         the y-axis and the corresponding concentration values at the
         x-axis. Can be used to check for measurement error variance
         heteroscedasticity. Often a supplement to \code{residual_plot}.
   \item \code{cv_vs_concentration: } Clinical sample coefficient of
         variation at the y-axis and the corresponding concentration values
         at the y-axis. Can be used to check whether a log-transformation
         would solve heteroscedasticity issues for the measurement error
         variance.
}

For formal testing, the end-user may use \code{perform_assessment_tests()}.

Note: In general, it is not recommended to specify
\code{additional_arguments}. The default settings are usually adequate in
most scenarios.
}
\examples{
# Required packages
library(smooth.commutability)
library(data.table)

# Read example data from smooth.commutability package
test_cs_data <- copy(crp_cs_data)

# Draw Residual Plots Using Weighted Smoothing Splines
plot_assessment_plots(data = test_cs_data,
                      method = "ssw",
                      type = "residual_plot",
                      draw_curves = TRUE,
                      plot_theme = "noklus")

# Draw QQ Normal Plots Using Deming Regression
plot_assessment_plots(data = test_cs_data,
                      method = "fg",
                      type = "qq_plot",
                      draw_curves = FALSE,
                      plot_theme = "depression")

# Draw Coefficient of Variation Versus Concentration Plots
plot_assessment_plots(data = test_cs_data,
                      type = "cv_vs_concentration",
                      draw_curves = TRUE,
                      plot_theme = "soft",
                      additional_arguments = list(cv_percent = TRUE))


}
