% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_commutability_evaluation_plots.R
\name{plot_commutability_evaluation_plots}
\alias{plot_commutability_evaluation_plots}
\title{Plots commutability evaluation plots based on ce_data, pb_data and ce_data}
\usage{
plot_commutability_evaluation_plots(cs_data, pb_data, ce_data)
}
\arguments{
\item{cs_data}{\code{list}, \code{data table} or \code{data frame} - Data containing clinical samples' measurements}

\item{pb_data}{\code{list}, \code{data table} or \code{data frame} - Data containing prediction band data}

\item{ce_data}{\code{list}, \code{data table} or \code{data frame} - Data containing prediction evaluation data for evaluated control materials}
}
\value{
\code{ggplot2} object that is a grid of plots having dimensions corresponding to the number of unique IVD-MD comparisons
}
\description{
Plots commutability evaluation plots based on ce_data, pb_data and ce_data
}
\examples{
print(1)
}
