% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_commutability_evaluation_plots.R
\name{plot_commutability_evaluation_plots}
\alias{plot_commutability_evaluation_plots}
\title{Plots commutability evaluation plots based on ce_data, pb_data and ce_data}
\usage{
plot_commutability_evaluation_plots(
  cs_data,
  pb_data,
  ce_data,
  exclude_rings = FALSE,
  exclude_cs = FALSE,
  plot_theme = c("custom", "default", "noklus", "soft", "depression", "happy"),
  additional_arguments = NULL,
  testing = FALSE
)
}
\arguments{
\item{cs_data}{This is either a \code{list}, \code{data.table} or \code{data.frame} containing mean-of-replicates clinical sample data. Data should be grouped by both \code{comparison} and \code{SampleID}.}

\item{pb_data}{This can be a \code{list}, \code{data.table} or \code{data.frame} containing pointwise prediction interval or prediction band data. This data should be grouped by \code{comparison}.}

\item{ce_data}{This should be a \code{list}, \code{data.table} or \code{data.frame} that holds prediction interval data along with commutability evaluation data for assessed external quality assessment materials or reference materials. It must be grouped by both \code{comparison} and \code{SampleID}.}

\item{exclude_rings}{A \code{logical} value that indicates whether circles, denoting the relative consistency of commutability evaluation conclusions, should be omitted. The default is \code{FALSE}.}

\item{exclude_cs}{A \code{logical} value that determines whether to exclude clinical sample data from the plots. The default is \code{FALSE}.}

\item{plot_theme}{A \code{character} string specifying the desired plotting theme. Options include 'custom', 'default', 'noklus', 'soft', 'depression', and 'happy'. The default is 'custom', which applies no particular theme to the plots unless specified in \code{additional_arguments}.}

\item{additional_arguments}{This is a \code{list} specifying additional arguments to modify the output plot. The default value is \code{NULL}, indicating that the plot will be generated using the default settings. Optional additional arguments can be provided to customize the plot according to specific requirements.
\itemize{
  \item{\code{main_title}: }{Main title of the plot}
  \item{\code{sub_title}: }{Sub title of the plot}
  \item{\code{x_name}: }{Title of the x-axis}
  \item{\code{y_name}: }{Title of the y-axis}
  \item{\code{n_breaks}: }{Number of axis-breaks}
  \item{\code{title_color}: }{Text color of 'main_title'. Default color is \code{'black'}}
  \item{\code{sub_title_color}: }{Text color of 'sub_title'. Default color is \code{'black'}}
  \item{\code{pb_fill:  }}{Fill color of prediction bands. Default fill color is \code{'green'}}
  \item{\code{pb_border}:  }{Color of the border of the prediction bands. Default color is \code{'black'}}
  \item{\code{curve}:  }{Should a curve be included to the clinical samples measurements. Default is \code{FALSE}}
  \item{\code{curve_type}: }{Which type of curve should be included. Only relevant if \code{curve = TRUE}. Possible values are 'equivalence_curve', 'fitted_curve' and 'flexible_curve'}
  \item{\code{curve_color}: }{Color of curve. Only relevant if \code{curve = TRUE}. Default is \code{'black'}.}
  \item{\code{point_shape}: }{Shape of clinical sample points. Possible inputs are 'circle' (default), 'square', 'diamond' and 'triangle'}
  \item{\code{point_size}: }{Size of clinical sample points. Possible inputs are of type \code{double} and must larger than 0. Default is automatically determined}
  \item{\code{point_fill}: }{fill color of clinical sample points. May be a color defined in \code{colors()} or a HEX code}
  \item{\code{point_border}: }{border color of clinical sample points. May be a color defined in \code{colors()} or a HEX code. Default color is \code{'black'}}
  \item{\code{comparison_fill}: }{Fill color for the individual plot labels listing the IVD-MD comparisons. Must be a valid color name or HEX code}
  \item{\code{comparison_text_color}: }{Text color for the individual plot labels listing the IVD-MD comparisons. Must be a valid color name or HEX code}
  \item{\code{hide_prediction_intervals}: }{Should the black prediction intervals for each EQAM be hidden?}
  \item{\code{legend_fill}: }{Fill color for the legend regions. Must be a valid color name or HEX code}
  \item{\code{legend_text_color}: }{Text color for the legends. Must be a valid color name or HEX code}
}}

\item{testing}{A \code{logical} value indicating whether the function output should be test-friendly. This parameter is primarily intended for maintainers and not typically modified by end-users. The default value is \code{FALSE}.}
}
\value{
A \code{ggplot2} object that is a grid of plots having dimensions corresponding to the number of unique IVD-MD comparisons
}
\description{
Plots commutability evaluation plots based on ce_data, pb_data and ce_data
}
\examples{
print(1)
}
