% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{repair_data}
\alias{repair_data}
\title{Repair a Method Comparison Dataset}
\usage{
repair_data(
  data,
  type = "cs",
  remove_invalid_methods = TRUE,
  include_repair_summary = FALSE
)
}
\arguments{
\item{data}{A \code{data.table}, \code{list} or \code{data.frame}. The data
that is desired to repair.}

\item{type}{A \code{character} string. The type of \code{data} must be
one of the following:
\itemize{
   \item \code{'cs': } \code{data} is treated as clinical sample
         data.
   \item \code{'eqam': } \code{data} is treated as evaluated
         material data.
}
The suitability requirements for \code{data} will depend on
\code{type}. For example, there are different requirements for
clinical sample data and external quality assessment material
data.}

\item{remove_invalid_methods}{A \code{logical} value. If \code{TRUE}
(default), IVD-MDs deemed invalid will be
removed during the repair process.}

\item{include_repair_summary}{A \code{logical} value. If \code{TRUE}, a
summary of the repair process will be included
in the output.}
}
\value{
If \code{include_repair_summary} is \code{FALSE}, returns a repaired
\code{data.table}. If \code{TRUE}, returns a \code{list} containing:
\itemize{
 \item \code{repaired_data: } A \code{data.table}. The repaired \code{data}.
 \item \code{repair_summary: } A \code{data.table} summarizing the repair
       process.
}
}
\description{
This function attempts to repair and standardize the input dataset for use
in method comparison analyses. It performs various checks and corrections to
ensure data integrity and compatibility with other functions in the package.
}
\details{
The function performs the following steps:
\enumerate{
 \item Checks the input data using the \code{check_data()} function.
 \item Converts the data to a \code{data.table} if possible.
 \item Standardizes ID column names: \code{SampleID} and \code{ReplicateID}.
 \item Converts methods to \code{numeric} format if necessary.
 \item Reorders \code{numeric} columns alphabetically.
 \item Removes invalid \code{numeric} columns if specified.
 \item Removes rows with \code{NA} values in ID columns.
 \item Sorts the data by \code{SampleID} and then \code{ReplicateID}.
 \item Converts \code{SampleID} and \code{ReplicateID} to \code{character}.
}
}
\examples{
# Required packages
library(smooth.commutability)
library(data.table)

# Get example data from smooth.commutability package
test_cs_data <- reverse_comparison_data(copy(crp_cs_data))
test_eq_data <- reverse_comparison_data(copy(crp_eqam_data))

# Repaired clinical sample data (with repair summary)
repaired_test_cs_data <- repair_data(
 data = test_cs_data,
 type = "cs",
 remove_invalid_methods = TRUE,
 include_repair_summary = TRUE
)

# The repaired data:
repaired_test_cs_data$repaired_data

# the repair summary:
repaired_test_cs_data$repair_summary

# Repaired external quality assessment material data (without repair summary)
repaired_test_eq_data <- repair_data(
 data = test_eq_data,
 type = "eqam",
 remove_invalid_methods = TRUE,
 include_repair_summary = TRUE
)

# The repaired data
repaired_test_eq_data

}
