% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_zetas.R
\name{simulate_zeta}
\alias{simulate_zeta}
\title{Simulate a Point Estimate of \eqn{\zeta}}
\usage{
simulate_zeta(parameters, simulation_indices)
}
\arguments{
\item{parameters}{A \code{list}, \code{data.table}, or \code{data.frame}.
The simulation parameters. See \code{?sim_eqa_data()} for
available options and general guidance.}

\item{simulation_indices}{A \code{numeric} vector. Kept available for the
\code{boot()} method. See the \code{boot-package}.}
}
\value{
A \code{double}. The simulated \eqn{\hat{\zeta}} value.
}
\description{
This function simulates \eqn{\hat{\zeta}} once based on \code{parameters}.
It is almost exclusively used internally within the more comprehensive
\code{simulate_zetas()} function.
}
\details{
Given the values found in \code{parameters}, one value of \eqn{\hat{\zeta}}
is simulated. This is typically not a function end-users should use directly
but rather a helper-function for the considerably more comprehensive and
useful \code{simulate_zetas()}. The second argument
\code{simulation_indices} have no actual function. It is required because
the \code{boot-package} requires it in its methods but it is never actually
used.
}
\examples{
# Simulate a zeta value for a pair of IVD-MDs susceptible to random DINS
simulate_zeta(parameters = list(n = 25, R = 3, prop = 0.05, mmax = 5))

# Simulate a zeta value for a pair of IVD-MDs susceptible to upper systematic DINS
simulate_zeta(parameters = list(n = 20, R = 4, qpos = 1, qran = 0.2, mmax = 5))
}
