% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_zetas.R
\name{simulate_zetas}
\alias{simulate_zetas}
\title{Monte Carlo Simulations of \eqn{\hat{\zeta}}}
\usage{
simulate_zetas(
  n,
  parameters,
  statistic = median,
  m = 1L,
  attach = TRUE,
  probs = 0.99,
  simplify = FALSE
)
}
\arguments{
\item{n}{An \code{integer}. Must be larger than or equal to \code{1}. The
desired number of simulated \eqn{\hat{\zeta}} values for each
replication of the target \code{statistic}(s).}

\item{parameters}{A \code{list}, \code{data.table}, or \code{data.frame}.
The simulation parameters. Refer to
\code{?sim_eqa_data} for more information.}

\item{statistic}{A \code{function} or \code{character} string. The desired
                 statistic to approximate using the \code{n} simulated
                 \eqn{\hat{\zeta}} values. If passed as a \code{character}
                 string, it must be one of the following:
\itemize{
  \item \code{'none': } Returns simulated samples of \eqn{\hat{\zeta}}
        values. Thus, no statistic is calculated.
  \item \code{'all': } Approximates the \code{mean}, \code{variance},
        \code{skewness}, \code{kurtosis}, and the following percentiles:
        1st, 2.5th, 5th, 10th, 25th, 50th, 75th, 90th, 95th, 97.5th, and 99th.
  \item \code{'all2': } Approximates the \code{mean}, \code{variance},
        \code{skewness}, \code{kurtosis}, and the 1st and 99th percentiles.
  \item \code{'moments': } Approximates the \code{mean}, \code{variance},
        \code{skewness} and \code{kurtosis}.
  \item \code{'moments12': } Approximates the \code{mean} and
        \code{variance}.
  \item \code{'moments34': } Approximates the \code{skewness} and
        \code{kurtosis}.
  \item \code{'quantiles': } Approximates the following percentiles: 1st,
        2.5th, 5th, 10th, 25th, 50th, 75th, 90th, 95th, 97.5th, and 99th.
  \item \code{'quantiles2': } Approximates the following percentiles: 1st,
        25th, 50th, 75th, and 99th.
  \item \code{'quartiles': } Approximates the three quartiles: the 25th,
        50th, and 75th percentiles.
}}

\item{m}{An \code{integer} \eqn{\geq 1}. The number of replicated
approximations of \code{statistic}. It is not advisable to let
\code{m} \eqn{\times} \code{n} \eqn{\geq 1e7}, due to unacceptable
computation times.}

\item{attach}{A non-missing \code{logical} value. If \code{TRUE} (default),
the approximated statistic(s) are merged with
\code{parameters} allowing a direct mapping between simulation
parameters and approximated statistic(s).}

\item{probs}{A \code{double} \eqn{\in} \code{[0, 1]}. Only relevant if
\code{statistic = quantile(...)}.}

\item{simplify}{A non-missing \code{logical} value. If \code{TRUE}, the
output is attempted to be converted into a simplified
format. Typically results in a \code{data.table} object
or a \code{vector}, depending on \code{attach}.}
}
\value{
A \code{list}, \code{data.table} or \code{numeric} vector.
}
\description{
Simulate statistics or a sample of \eqn{\hat{\zeta}} values. In practice,
\code{n} \eqn{\hat{\zeta}} values are simulated \eqn{m} times, and for each
of the \code{m} samples, an individual or a set of statistics are estimated.
}
\examples{
# Required packages
library(fasteqa)
library(data.table)

# Approximate the standard deviation of the point estimate of zeta

# Simulation parameters (data with heteroscedasticity)
sim_pars <- list(n = 20,
                 R = 4,
                 cvx = 0.01,
                 cvy = 0.02,
                 cil = 5,
                 ciu = 90,
                 eta = 5,
                 eta0 = 1)

# Simulated standard deviation based on 100 simulated point estimates of zeta
sd_zeta_hat <- simulate_zetas(n = 100,
                              parameters = sim_pars,
                              statistic = sd,
                              attach = TRUE)

print(sd_zeta_hat)
}
