% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Transformation of EQA data}
\usage{
transform_data(data, transformation = "log")
}
\arguments{
\item{data}{A \code{list}, \code{data table} or \code{data frame} that must contain \code{MP_A} and \code{MP_B}}

\item{transformation}{\code{character} - Either a string seperated by space or symbol so that we have transformation|seperator|value. For example, if the cube root is required, you e.g., write root#3 or root_3. Possible transformations are:
\itemize{
  \item{\code{log}: }{log-transformation, with values representing the base of the log-transformation. All values to be transformed must be positive}
  \item{\code{root}: }{root-transforomation, with values representing which root is taken, e.g., 2 for square root, 3 for cube root or other integers. All values to be transformed values must be non-negative}
  \item{\code{pow}: }{power-transformation, with values representing the power taken, e.g., 2 for squaring, 3 for cubing or other integers. If value < 1, all values to be transformed must be non-negative}
  \item{\code{inv}: }{inverse-transformation, with values representing the power of the denominator of the invserse transformation. If value < 1, all values transformed to be transformed must be non-negative}
  \item{\code{id}: }{indentity-transformation, which means no transformation at all}
}}
}
\value{
The transformed \code{data} as a \code{data table}
}
\description{
Transformation of EQA data
}
\examples{
print(1)
}
