% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_checks.R
\name{validate_data}
\alias{validate_data}
\title{Validates and Converts A \code{data} Argument}
\usage{
validate_data(
  data,
  argument_name = "data",
  required_columns = c("comparison", "SampleID", "ReplicateID", "MP_A", "MP_B")
)
}
\arguments{
\item{data}{A \code{data.table}, \code{list} or \code{data.frame}. Must
contain \code{required_columns}.}

\item{argument_name}{A \code{character} string. The name of the argument.}

\item{required_columns}{A \code{character} vector. The required columns in
\code{data}.}
}
\value{
A \code{data.table}. The validated and converted \code{data}.
}
\description{
Validates and converts \code{data} to a \code{data.table} object.
}
\details{
When \code{data} fail a particular validation test, an error is thrown that
describes why \code{data} does not pass as a valid \code{data} argument.
}
\examples{
print(1)
}
